### 2022-12-01 #4986 @aholmes
> I would appreciate any suggestions for potential reviewers. Please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Hello @danielskatz, I'm interested in reviewing.
