### 2022-12-02 #4986 @yguclu
After a keyword search in [this list of people](https://bit.ly/joss-reviewers) I can suggest the following reviewers:
- jeremylt
- gassmoeller
- svchb
- victorapm
- kris-rowe

I do not know any of them personally, but I hope they can find our software of interest.
