### 2022-12-02 #4986 @danielskatz
👋 @jeremylt & @gassmoeller - Would either of you be willing to review this JOSS submission?
