### 2022-12-03 #4986 @boegel
@danielskatz I'm interested in reviewing (first time for JOSS)
