### 2022-12-03 #4991 @editorialbot
[REVIEW]: Pyccel: a Python-to-X transpiler for scientific high-performance computing
**Submitting author:** @yguclu (<a href="http://orcid.org/0000-0003-2619-5152">Yaman Güçlü</a>)
**Repository:** https://github.com/pyccel/pyccel/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.7.2
