+++ pyccel/ast/cudaext.py
@@ -203,33 +203,9 @@ class CudaGrid(PyccelAstNode)               :
             return expr[0]
         return PythonTuple(*expr)
 
-class CudaToDevice(CudaArray):
-    """Represents a call to cuda.to_device for code generation.
-        arg : Variable
-    """
-    def __init__(self, arg):
-        if not isinstance(arg, Variable) or not arg.is_ndarray:
-            raise TypeError("Argument must be an ndarray variable")
-        order = arg.order
-        current_context =  'device' if not arg.memory_location == 'host' else arg.memory_location
-        super().__init__(arg, order=order, current_context=current_context, memory_location='device')
-
-class CudaToHost(CudaArray):
-    def __init__(self, arg):
-        """Represents a call to cuda.to_host for code generation.
-            arg : Variable
-        """
-        if not isinstance(arg, Variable) or not arg.is_ndarray:
-            raise TypeError("Argument must be an ndarray variable")
-        order = arg.order
-        current_context = 'device' if not arg.memory_location == 'host' else arg.memory_location
-        super().__init__(arg, order=order, current_context=current_context, memory_location='host')
-
 cuda_funcs = {
     # 'deviceSynchronize' : CudaDeviceSynchronize,
     'array'             : PyccelFunctionDef('array'             , CudaArray),
-    'to_device'         : PyccelFunctionDef('to_device'         , CudaToDevice),
-    'to_host'           : PyccelFunctionDef('to_host'           , CudaToHost),
     'deviceSynchronize' : PyccelFunctionDef('deviceSynchronize' , CudaDeviceSynchronize),
     'threadIdx'         : PyccelFunctionDef('threadIdx'         , CudaThreadIdx),
     'blockDim'          : PyccelFunctionDef('blockDim'          , CudaBlockDim),
