### 2022-12-04 pyccel/pyccel#1243 @aholmes
[JOSS] Contribution and authorship
Hello,

Regarding openjournals/joss-reviews#4991 - on reviewing "contribution and authorship" I noticed this repository has 24 contributors, while the paper has four authors listed. Is this correct?
