### 2022-12-04 ed01ab7
Add `pytest` dependencies to setup.cfg.
JOSS submission [requires](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) "... that software should be feature-complete (i.e.,
no half-baked solutions), packaged appropriately according to common
community standards for the programming language being used." I
identified that, upon installing the defined dependencies (in setup.cfg) with `pip install -e .`,
`pytest` was not runnable. The dependencies for `pytest` are listed in [README.md](https://github.com/pyccel/pyccel/blob/8ca23f07fa9aed4ed1974c7b7d6f427bcce7f7df/REA
[…]
