### 2022-12-04 pyccel/pyccel#1245 @aholmes
[JOSS] Community guidelines - add a CONTRIBUTING.md file
Regarding openjournals/joss-reviews#4991 - "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I recommend adding a [CONTRIBUTING.md](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file to the repository to outline how others might contribute to this repository.
