### 2022-12-05 4f92892
forced ndarray creation with numpy.array to remove confusion with normal list
