### 2022-12-05 pyccel/pyccel#1249 @aholmes
[JOSS] Reproducibility - consider locking dependency versions
Regarding openjournals/joss-reviews#4991

Hello,

One aspect of reproducibility involves the dependencies used to make any software. Dependencies change often from one release to another, sometimes dramatically. When changes occur that change the behavior in their dependents, the dependent software is then broken. For any software, it is important that the versions of dependencies for a given release of the software are the same for all new installations and usages of the software - and it is especially important for scientific software and papers in order to allow for peer review and equivalent results in further research.

An example where this could be problematic is in the specification of `numpy` as a dependency. In version [v1.7.0](https://github.com/pyccel/pyccel/blob/v1.7.0/setup.cfg#L16), for example, no version is specified. This means the latest version of `numpy` will always be used, which might cause the release v1.7.0 to fail to work correctly - in the best case causing errors, and worst case appearing to work correctly, but producing erroneous results.

The fix is to specify either exact versions of dependencies, or ranges of dependencies known to work (and that have been tested). While [the documentation](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html) on ways to do this is a little obtuse, this can all be handled in setup.cfg. For example, `sympy` can be specified as

```ini
[options]
install_requires =
    sympy >= 1.2, <= 1.7.1
```
