### 2022-12-06 pyccel/pyccel#1243 @yguclu
Hi, the paper was written by the project creators, maintainers, and main contributors over several years of work. These are the 4 authors of the paper.

We try to keep an up-to-date list with the names of every person who has made significant contributions to the Pyccel library in the [AUTHORS](https://github.com/pyccel/pyccel/blob/master/AUTHORS) file, as is standard practice for a large open-source project.

In the paper we thank all contributors in the **Acknowledgment** section.
