### 2022-12-06 pyccel/pyccel#1245 @yguclu
Thanks for pointing this out. We are working on it.
