+++ .github/workflows/Github_pytest.yml
@@ -5,45 +5,45 @@ on:
     branches: [ master ]
 
 jobs:
-  #Linux:
+  Linux:
 
-  #  runs-on: ubuntu-latest
+    runs-on: ubuntu-latest
 
-  #  steps:
-  #    - uses: actions/checkout@v3
-  #    - name: Set up Python 3.7
-  #      uses: actions/setup-python@v4
-  #      with:
-  #        python-version: 3.7
-  #    - name: Install dependencies
-  #      uses: ./.github/actions/linux_install
-  #    - name: Install python dependencies
-  #      uses: ./.github/actions/pip_installation
-  #    - name: Coverage install
-  #      uses: ./.github/actions/coverage_install
-  #    - name: Fortran/C tests with pytest
-  #      uses: ./.github/actions/pytest_run
-  #    - name: Python tests with pytest
-  #      uses: ./.github/actions/pytest_run_python
-  #    - name: Parallel tests with pytest
-  #      uses: ./.github/actions/pytest_parallel
-  #    - name: Test with valgrind for memory leaks
-  #      uses: ./.github/actions/valgrind_run
-  #    - name: Collect coverage information
-  #      continue-on-error: True
-  #      uses: ./.github/actions/coverage_collection
-  #    - name: Run codacy-coverage-reporter
-  #      uses: codacy/codacy-coverage-reporter-action@master
-  #      continue-on-error: True
-  #      with:
-  #        project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
-  #        coverage-reports: cobertura.xml
-  #    - name: Save code coverage report
-  #      uses: actions/upload-artifact@v3
-  #      with:
-  #        name: coverage-artifact
-  #        path: cobertura.xml
-  #        retention-days: 1
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.7
+      - name: Install dependencies
+        uses: ./.github/actions/linux_install
+      - name: Install python dependencies
+        uses: ./.github/actions/pip_installation
+      - name: Coverage install
+        uses: ./.github/actions/coverage_install
+      #- name: Fortran/C tests with pytest
+      #  uses: ./.github/actions/pytest_run
+      - name: Python tests with pytest
+        uses: ./.github/actions/pytest_run_python
+      - name: Parallel tests with pytest
+        uses: ./.github/actions/pytest_parallel
+      - name: Test with valgrind for memory leaks
+        uses: ./.github/actions/valgrind_run
+      - name: Collect coverage information
+        continue-on-error: True
+        uses: ./.github/actions/coverage_collection
+      - name: Run codacy-coverage-reporter
+        uses: codacy/codacy-coverage-reporter-action@master
+        continue-on-error: True
+        with:
+          project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
+          coverage-reports: cobertura.xml
+      - name: Save code coverage report
+        uses: actions/upload-artifact@v3
+        with:
+          name: coverage-artifact
+          path: cobertura.xml
+          retention-days: 1
 
   #Windows:
 
