### 2022-12-06 da1696f
Use PR dictionary description
