### 2022-12-06 7dccc93
Add additional issue type
