+++ .github/ISSUE_TEMPLATE/installation-problem.md
@@ -1,0 +1,24 @@
+---
+name: Installation problem
+about: Report an installation problem
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Relevant Discussion**
+
+Most installation problems are not directly caused by pyccel, but are due to dependencies. Please link to the discussion in the [discussion tab](https://github.com/pyccel/pyccel/discussions/categories/q-a) where the problem was investigated.
+
+**Description**
+
+A clear and concise description of the problem encountered.
+
+**Environment information**
+
+Python version :
+
+OS version :
+
+Versions for any relevant dependencies:
