+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -7,6 +7,12 @@ assignees: ''
 
 ---
 
+**Relevant Discussion**
+
+Non-trivial features require a lot of discussion to find the best implementation in the target languages.
+This discussion should take place in [discussion tab](https://github.com/pyccel/pyccel/discussions/categories/q-a).
+Please link to the discussion where the proposed feature was discussed.
+
 **Describe the feature**
 A clear and concise description of what you would like to be implemented.
 
@@ -16,5 +22,5 @@ Provide code which does not currently work but which should do when this issue i
 Code here
 ```
 
-**Language**
-If appropriate, please specify which language the python code is translated to (Fortran by default)
+**Proposed Solution**
+Briefly summarise how you expect this feature to be implemented
