### 2022-12-06 pyccel/pyccel#1249 @aholmes
> Our unit tests run on a daily basis using the latest available versions of our dependencies, hence if we catch a problem with any of them we immediately release a new (minor) version of Pyccel where we specify the maximum version of the problematic library.

This is great and definitely helps with any future versions and downloads of the software, however, users of previous versions will need to be made aware of any new issues and future updates, and we (package maintainers in general) aren't able to dictate how users of our software will manage the dependency on our own software - meaning they may end up using a version of pyccel that, when `pip install .` is run, pulls a version of a dependency that is no longer compatible with that version of pyccel. In cases where this causes the software to loudly fail (errors and such), users can quickly discover what is going wrong. In cases of silent failure (subtle changes in behavior, etc.), users may never discover that something isn't correct.

It's this case in particular that locking versions helps with - for users on past versions of the software who may or may not be monitoring updates.

> I understand the need for reproducibility in science, but I think that this could be better served by using `pip freeze > requirements.txt`

This is a reasonable compromise. Can I propose that the "Pyccel Installation Methods" section in README.md is updated to reflect this, if this method is pursued? It may make sense to categorize "user installation methods" using requirements.txt, and "developer installation methods" using `pip install -e .`. I also suggest adding a note about using a virtual environment to install and use the software to help avoid the case where a user unintentionally changes their system Python dependencies.

> Maybe we could be stricter in our dependencies when producing a Dockerfile

This is a great idea!

Another consideration would be setting specific version dependencies as part of the pyccel release process. setup.cfg can remain as is in the "main" branch, and versions can be specific in setup.cfg in a release branch. This way "main" isn't restricted, while the versions will be "self-contained," so to speak.
