### 2022-12-06 pyccel/pyccel#1249 @yguclu
Thanks for the clarifications! There is definitely much room for thought here... We'll get back to you ASAP
