+++ setup.cfg
@@ -22,6 +22,16 @@ python_requires = >= 3.7
 zip_safe = False
 include_package_data = True
 
+[options.extras_require]
+test =
+    pytest
+    scipy
+    mpi4py
+    coverage
+    astunparse
+    tblib
+    pytest-xdist
+
 [options.entry_points]
 console_scripts =
     pyccel = pyccel.commands.console:pyccel
