+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -7,10 +7,10 @@ assignees: ''
 
 ---
 
-**Describe the bug**
+### Describe the bug
 A clear and concise description of what the bug is.
 
-**To Reproduce**
+### To Reproduce
 Provide code to reproduce the behavior:
 ```python
 Code here
@@ -21,11 +21,11 @@ Provide the generated code, or the error message:
 Translated code here
 ```
 
-**Expected behavior**
+### Expected behavior
 A clear and concise description of what you expected to happen.
 
-**Language**
+### Language
 Please specify which language the python code is translated to (Fortran by default)
 
-**Additional context**
+### Additional context
 Add any other context about the problem here.
