+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -7,20 +7,20 @@ assignees: ''
 
 ---
 
-**Relevant Discussion**
+### Relevant Discussion
 
 Non-trivial features require a lot of discussion to find the best implementation in the target languages.
 This discussion should take place in [discussion tab](https://github.com/pyccel/pyccel/discussions/categories/q-a).
 Please link to the discussion where the proposed feature was discussed.
 
-**Describe the feature**
+### Describe the feature
 A clear and concise description of what you would like to be implemented.
 
-**Test Code**
+### Test Code
 Provide code which does not currently work but which should do when this issue is fixed:
 ```python
 Code here
 ```
 
-**Proposed Solution**
+### Proposed Solution
 Briefly summarise how you expect this feature to be implemented
