+++ .github/CONTRIBUTING.md
@@ -1,8 +1,11 @@
 # Contributing to pyccel
 
-We welcome any and all contributions to pyccel!
+Firstly, thanks for your interest in contributing code to pyccel!
 
-The following is a summary of the best ways to interact/help with this project.
+We welcome any and all contributions.
+
+There are many ways to help with the pyccel project which are more or less involved.
+The following is a summary of how you can do this.
 
 ## Asking questions
 
@@ -21,16 +24,12 @@ When reporting issues please include as much detail as possible about your
 operating system, and python version. Whenever possible, please
 also include a brief, self-contained code example that demonstrates the problem.
 
-## Ways to help
-
-There are many ways to help with the pyccel project which are more or less involved.
-
-### Answering questions
+## Answering questions
 
 Questions are sometimes asked in the [discussions tab](https://github.com/pyccel/pyccel/discussions/categories/q-a) or in issues.
 Any help answering these questions is appreciated as it helps our users find solutions as fast as possible.
 
-### Documentation
+## Documentation
 
 Reviewing the existing documentation is a great way to help out!
 We need to ensure that it is up-to-date and relevant, and that the examples are clear, relevant, and work correctly.
@@ -39,7 +38,7 @@ Missing documentation can also be added to improve the existing docs.
 
 Please open a pull request to add or change anything in the docs.
 
-### Reviewing a pull request
+## Reviewing a pull request
 
 There are many pull requests and reviewing them is a great way to help things get to the master branch faster.
 This is also a really good way to get to grips with the code base as you will see examples of many different areas of the code.
@@ -47,9 +46,7 @@ It's incredibly helpful to have pull requests go through as many reviews as poss
 As the clarity is subjective, more eyes can only improve the code base.
 The review process is described in the [developer docs](https://github.com/pyccel/pyccel/blob/master/developer_docs/review_process.md), keep an eye out for PRs tagged `needs_initial_review`.
 
-### Contributing code
-
-Firstly, thanks for your interest in contributing code to pyccel!
+## Contributing code
 
 Before contributing we strongly recommend you check out the [developer docs](https://github.com/pyccel/pyccel/tree/master/developer_docs).
 We try to flag `good-first-issue`s.
