### 2022-12-08 a223f1d
Improve bug template
