### 2022-12-09 pyccel/pyccel#1249 @yguclu
After the discussion in our developers' meeting, we plan to:
- Specify the minimum version of all our dependencies
- Clarify in the README how a `virtualenv` and `requirements.txt` file can be used for reproducibility
- Provide an optional `pyccel-test` command at installation, for running all tests (`pyccel-test all`) or just the basic ones (`pyccel-test basic`)
- Automatically store a `requirements-[version].txt` file every time we upload new benchmark results to the `pyccel-benchmarks` repository

@aholmes Does this sound as a good compromise to you?
