### 2022-12-09 pyccel/pyccel#1249 @aholmes
@yguclu Yes, I think this sounds reasonable.

One question - why `requirements-[version].txt` over just `requirements.txt`? This seems burdensome for you to maintain, and you would eventually have many of these files, if I understand correctly.
