### 2022-12-09 pyccel/pyccel#1249 @yguclu
For each release of Pyccel, we already generate the following files, which are added to the directory `version_specific_results/`:
- `pypi_performance_[python-version]_[pyccel-version].md`
- `pypi_performance_[python-version]_[pyccel-version]_compilation.png`
- `pypi_performance_[python-version]_[pyccel-version]_execution.png`

The idea is to also generate the corresponding requirements file by using
```bash
python3 -m pip freeze > pypi_performance_[python-version]_[pyccel-version]_requirements.txt
```

This way we will keep track of all the dependencies used to run the benchmarks.
