### 2022-12-09 pyccel/pyccel#1249 @yguclu
On a separate note, I am now realizing that adding all those PNG images to the git repository will cause it to grow in size very quickly, so it is not really sustainable in the long run. @EmilyBourne I think we need a better solution...
