### 2022-12-09 pyccel/pyccel#1249 @aholmes
> For each release of Pyccel, we already generate the following files

Okay, great!

> I am now realizing that adding all those PNG images

If I may suggest - SVG is a good alternative. The size is much smaller, and the changes can be diffed. GitHub has a really nice "SVG diff viewer" to visually see what's different.
