### 2022-12-11 pyccel/pyccel#1249 @EmilyBourne
> Provide an optional pyccel-test command at installation, for running all tests (pyccel-test all) or just the basic ones (pyccel-test basic)

I'm not sure how this should work given that we don't install the tests. Would it not make more sense to describe a command in the README?
