### 2022-12-11 d9804e9
Test minimum requirements
