+++ setup.cfg
@@ -13,23 +13,23 @@ long_description_content_type = text/markdown
 [options]
 packages = find:
 install_requires =
-    numpy >= 1.16
-    sympy >= 1.2
-    termcolor >= 1.0.0
-    textx >= 2.2
-    filelock >= 3.4.0
-python_requires = >= 3.7
+    numpy == 1.16
+    sympy == 1.2
+    termcolor == 1.0.0
+    textx == 2.2
+    filelock == 3.4.0
+python_requires = == 3.7
 zip_safe = False
 include_package_data = True
 
 [options.extras_require]
 test =
-    pytest >= 2.7
-    scipy >= 1.5.0
-    mpi4py >= 3.0.0
-    coverage >= 3.1
-    astunparse >= 1.6.0
-    pytest-xdist >= 1.16
+    pytest == 2.7
+    scipy == 1.5.0
+    mpi4py == 3.0.0
+    coverage == 3.1
+    astunparse == 1.6.0
+    pytest-xdist == 1.16
 
 [options.entry_points]
 console_scripts =
