+++ setup.cfg
@@ -18,7 +18,7 @@ install_requires =
     termcolor == 1.0.0
     textx == 2.2
     filelock == 3.4.0
-python_requires = == 3.7
+python_requires = >= 3.7
 zip_safe = False
 include_package_data = True
 
