+++ setup.cfg
@@ -13,7 +13,7 @@ long_description_content_type = text/markdown
 [options]
 packages = find:
 install_requires =
-    numpy == 1.16
+    numpy == 1.17
     sympy == 1.2
     termcolor == 1.0.0
     textx == 2.2
