+++ pyproject.toml
@@ -1,10 +1,10 @@
 [build-system]
 requires = [
     "setuptools >= 37, < 61",
-    "numpy == 1.20",
-    "sympy==1.2",
-    "termcolor == 1.0.0",
-    "textx==2.2",
-    "filelock == 3.4.0",
+    "numpy >= 1.16",
+    "sympy>=1.2",
+    "termcolor >= 1.0.0",
+    "textx>=2.2",
+    "filelock >= 3.4.0",
 ]
 build-backend = "setuptools.build_meta"
