### 2022-12-12 9ec1513
Community guidelines (#1252)
Improve the community experience by:
- Adding a CONTRIBUTING.md file
- Correcting the error in the feature issue template
- Adding new templates for documentation and installation problems

Fixes #1245 and #1206
