### 2022-12-12 d1e94b1
issue_1254 - Do not use exact comparison in tests involving floating point numbers
