### 2022-12-12 4e703de
Add minimum requirements (#1253)
Specify minimum version of dependencies. Fixes #1249. Minimum requirements were tested in commit [01bf26e](https://github.com/pyccel/pyccel/pull/1253/commits/01bf26e5a15915c2782428d531410d42251cffcc).
