### 2022-12-13 e03f7db
added is_list property to variable class
