### 2022-12-13 2747bfa
order check before transposing
