### 2022-12-14 d2ec6a8
Correct duplicate function with missing contents
