### 2022-12-15 5a38278
add support for cuda.random.randint
