### 2022-12-15 49571aa
Correct image type
