### 2022-12-15 9604278
simplified printing
