+++ pyccel/codegen/printing/ccode.py
@@ -933,7 +933,7 @@ class CCodePrinter(CodePrinter):
     def _print_PythonPrint(self, expr):
         self.add_import(c_imports['stdio'])
         self.add_import(c_imports['inttypes'])
-        _deallocate = []
+
         end = '\n'
         sep = ' '
         code = ''
@@ -984,20 +984,13 @@ class CCodePrinter(CodePrinter):
                 continue
             if isinstance(f, PythonType):
                 f = f.print_string
-
-            if isinstance(f, PythonList) or (isinstance(f, Variable) and f.cls_base == ListClass):
+            elif isinstance(f, PythonList) or (isinstance(f, Variable) and f.is_list):
                 if args_format:
                     code += formatted_args_to_printf(args_format, args, sep)
                     args_format = []
                     args = []
-                if isinstance(f, PythonList):
-                    var = self.scope.get_temporary_variable(f.dtype, rank=f.rank, memory_handling='heap', cls_base=ListClass)
-                    self._additional_code += self._print(Assign(var, f))
-                    code += f"print_list({self._print(var)}, 0);\n"
-                    _deallocate.append(Deallocate(var))
-                else:
-                    code += f"print_list({self._print(f)}, 0);\n"
-                code += formatted_args_to_printf(args_format, args, sep if (i+1 < len(orig_args)) else end)
+                new_line = '1' if i+1 == len(orig_args) else '0'
+                code += f"print_list({self._print(f)}, {new_line});\n"
             elif isinstance(f, FunctionCall) and isinstance(f.dtype, NativeTuple):
                 tmp_list = self.extract_function_call_results(f)
                 tmp_arg_format_list = []
@@ -1039,8 +1032,6 @@ class CCodePrinter(CodePrinter):
                 args.append(arg)
         if args_format:
             code += formatted_args_to_printf(args_format, args, end)
-        for arg in _deallocate:
-            code += self._print(arg)
         return code
 
     def find_in_dtype_registry(self, dtype, prec):
