### 2022-12-16 6accae5
updated _print_PythonList to use the function compount_stmt
