### 2022-12-16 282b761
updated _print_PythonListCount to print argument in the proper form
