### 2022-12-16 0419ba5
changed get_declare_type conditions for lists support
