### 2022-12-16 1cfcba7
added casting prefix for the return value of PythonListPop
