### 2022-12-16 cefe13b
updated _print_PyccelArraySize to get corresponding C expression for PythonList
