+++ pyccel/codegen/printing/ccode.py
@@ -1354,6 +1354,8 @@ class CCodePrinter(CodePrinter):
 
     def _print_PyccelArraySize(self, expr):
         arg    = expr.arg
+        if isinstance(arg, PythonList) or (isinstance(arg, Variable) and arg.is_list):
+            return '{}->size'.format(arg)
         if self.stored_in_c_pointer(arg):
             return '{}->shape[{}]'.format(self._print(ObjectAddress(arg)), self._print(expr.index))
         return '{}.shape[{}]'.format(self._print(arg), self._print(expr.index))
