### 2022-12-16 d78a6e4
calls fill_list inside _print_Assign to memcpy content of the list
