### 2022-12-20 ea08a0f
fix deallocate printer
