+++ pyccel/stdlib/cuda_ndarrays/cuda_ndarrays.h
@@ -24,4 +24,13 @@ void _cuda_array_fill_double(double c, t_ndarray arr);
 t_ndarray   cuda_array_create(int32_t nd, int64_t *shape, enum e_types type, bool is_view, enum e_memory_locations location);
 int32_t         cuda_free_array(t_ndarray dump);
 int32_t         cuda_free_pointer(t_ndarray dump);
+
+__host__ __device__
+int32_t cuda_free_host(t_ndarray arr);
+
+__host__ __device__
+int32_t cuda_free(t_ndarray arr);
+
+__host__ __device__
+int32_t cuda_free_pointer(t_ndarray arr);
 #endif
