### 2022-12-21 25c79d1
Change call to cudaDeviceSynchronize to cuda.synchronize()
