+++ tests/internal/scripts/ccuda/cuda_array_device.py
@@ -11,6 +11,6 @@ if __name__ == '__main__':
     threads_per_block = 5
     n_blocks = 1
     arr = cuda.array([0,1,2,3,4], memory_location = 'device')
-    cuda.deviceSynchronize()
+    cuda.synchronize()
     square[n_blocks, threads_per_block](arr)
-    cuda.deviceSynchronize()
+    cuda.synchronize()
