### 2022-12-21 1bc9f61
issue_1254 - updated the tests that use exact comparison to use approximate comparisons for floats
