+++ tests/epyccel/test_epyccel_complex_func.py
@@ -7,6 +7,9 @@ from numpy.random import rand, randint
 import modules.complex_func as mod
 from pyccel.epyccel import epyccel
 
+ATOL = 1e-15
+RTOL = 2e-14
+
 @pytest.mark.parametrize("f", [ mod.create_complex_literal__int_int,
                            mod.create_complex_literal__int_float,
                            mod.create_complex_literal__int_complex,
@@ -43,7 +46,7 @@ def test_create_complex_var__complex_float(language):
 
     a = complex(randint(100), randint(100))
     b = rand()*100
-    assert f_epyc(a,b) == f(a,b)
+    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
 
 def test_create_complex_var__complex_complex(language):
     f = mod.create_complex_var__complex_complex
@@ -51,7 +54,7 @@ def test_create_complex_var__complex_complex(language):
 
     a = complex(randint(100), randint(100))
     b = complex(randint(100), randint(100))
-    assert f_epyc(a,b) == f(a,b)
+    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
 
 def test_create_complex__int_int(language):
     f = mod.create_complex__int_int
@@ -72,21 +75,21 @@ def test_create_complex__float_float(language):
     f_epyc = epyccel(f, language = language)
 
     a = rand()*100
-    assert f_epyc(a) == f(a)
+    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_create_complex_0__float_float(language):
     f = mod.create_complex_0__float_float
     f_epyc = epyccel(f, language = language)
 
     a = rand()*100
-    assert f_epyc(a) == f(a)
+    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_create_complex__complex_complex(language):
     f = mod.create_complex__complex_complex
     f_epyc = epyccel(f, language = language)
 
     a = complex(randint(100), randint(100))
-    assert f_epyc(a) == f(a)
+    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_cast_complex_1(language):
     f = mod.cast_complex_1
@@ -100,7 +103,7 @@ def test_cast_complex_2(language):
     f_epyc = epyccel(f, language = language)
 
     a = np.complex128(complex(randint(100), randint(100)))
-    assert f_epyc(a) == f(a)
+    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_cast_float_complex(language):
     f = mod.cast_float_complex
@@ -108,4 +111,4 @@ def test_cast_float_complex(language):
 
     a = rand()*100
     b = complex(randint(100), randint(100))
-    assert f_epyc(a,b) == f(a,b)
+    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
