### 2022-12-22 b9cff43
remove decorator from cuda_free_host and add test
