+++ tests/internal/scripts/ccuda/deallocate.py
@@ -1,10 +1,12 @@
-from pyccel.decorators import kernel, types
+from pyccel.decorators import kernel
 from pyccel import cuda
-import cupy as cp
 import math
 
 @kernel
 def func(arr:'int[:]'):
+    """
+    Takes an array 'arr' as input and raise each element to the power 2.
+    """
     i = cuda.grid(0)
     arr[i] = math.pow(arr[i], 2)
 
