### 2022-12-26 e4b3ac0
Deactivate deprecated tests
