### 2022-12-26 3b55eaa
Use correct bounds to avoid runtime warning
