+++ developer_docs/order_docs.md
@@ -2,28 +2,12 @@
 
 ## Order in numpy
 
-`order` is the parameter given to the `numpy.array` function in order to choose how the array is stored in memory.  
-Both of the orders discussed here (`order_f` and `order_c`) are stored **contiguously** in memory, but they differ in how they are arranged.
-
-### Order F
-
-`order='F'` tells `Numpy` to store the array column by column (column-major), example:  
-
-```python
-import numpy as np
-
-if __name__ == "__main__":
-  a = np.array([[1, 2, 3],
-                [4, 5, 6],
-                [7, 8, 9]], order='F')
-  print(a.ravel('K'))
-```  
-
-`array.ravel('k')` shows us how the array is stored in memroy, this python script will output `[1 4 7 2 5 8 3 6 9]`, notice that the columns are stored one after the other.  
+`order` is the parameter given to the `numpy.array` function in order to choose how the array is stored in memory.
+Both of the orders discussed here (`order_c` and `order_f`) are stored **contiguously** in memory, but they differ in how they are arranged.
 
 ### Order C
 
-`order='C'` on the other hand tells `Numpy` to store the array row by row (row-major), example:  
+`order='C'` tells `Numpy` to store the array row by row (row-major), example:
 
 ```python
 import numpy as np
@@ -33,13 +17,31 @@ if __name__ == "__main__":
                 [4, 5, 6],
                 [7, 8, 9]], order='C') # order='C' is the default in numpy.array
   print(a.ravel('K'))
- ```  
- 
+ ```
+
+`array.ravel('k')` shows us how the array is stored in memory.
 This python script will output `[1 2 3 4 5 6 7 8 9]`, notice that the rows are stored one after the other.
+This is the default behaviour in python.
+
+### Order F
+
+`order='F'` on the other hand tells `Numpy` to store the array column by column (column-major), example:
+
+```python
+import numpy as np
+
+if __name__ == "__main__":
+  a = np.array([[1, 2, 3],
+                [4, 5, 6],
+                [7, 8, 9]], order='F')
+  print(a.ravel('K'))
+```
+
+This python script will output `[1 4 7 2 5 8 3 6 9]`, notice that the columns are stored one after the other.
 
 ### Printing and indexing in `Numpy`
 
-`order` in `Numpy` does not affect the indexing or the printing, unlike `transposing`, the `shape` of the array remains the same, only the `strides` change, example:  
+`order` in `Numpy` does not affect the indexing or the printing, unlike `transposing`, the `shape` of the array remains the same, only the `strides` change, example:
 
 ```python
 import numpy as np
@@ -56,39 +58,39 @@ if __name__ == "__main__":
    print(a)
    # output:[[1 2]
    #         [4 5]
-   #         [7 8]] 
+   #         [7 8]]
    print(b)
    # output:[[1 2]
    #         [4 5]
    #         [7 8]]
-   
+
    print(a[2][1], a[0][0], a[1]) # output: 8 1 [4 5]
    print(b[2][1], b[0][0], b[1]) # output: 8 1 [4 5]
-```  
+```
 
-`arr.strides` is the variable that helps us navigate the array (indexing/printing), it tells us how many bytes we have to skip in memory to move to the next position along a certain axis (dimension). For example for `memory_layout_of_a = [1 4 7 2 5 8]` and `strides_of_a = (8, 24)`, we have to skip 8 bytes (1 element for `int64`) to move to the next row, but 24 bytes (3 elements for `int64`) to get to the same position in the next column of `a`.  
[…]
