### 2022-12-26 d7de918
replaces 'string' with 'array' and `i, j' with 'row, column'
not all 'i, j' are replaced, need to understand something in fortran docs
