### 2022-12-26 1d6a5d7
mentions that one dimensional arrays need no order
