### 2022-12-26 fb9ccfd
issue_1254 - remove unnecessary approximate comparisons for floats
Correct image type (#1262)

The type of the figures in the pyccel-benchmark repository was changed from .png to .svg which broke pyccel's README. This PR corrects the extension and fixes #1261.
