+++ tests/epyccel/test_epyccel_python_annotations.py
@@ -5,9 +5,6 @@ from numpy.random import randint
 from pyccel.epyccel import epyccel
 from modules import python_annotations
 
-RTOL = 2e-14
-ATOL = 1e-15
-
 def test_array_int32_1d_scalar_add(language):
 
     f1 = python_annotations.array_int32_1d_scalar_add
@@ -76,7 +73,7 @@ def test_array_real_1d_scalar_add(language):
     f1(x1, a)
     f2(x2, a)
 
-    assert np.allclose(x1, x2, rtol=RTOL, atol=ATOL)
+    assert np.array_equal( x1, x2 )
 
 
 def test_array_real_2d_F_scalar_add(language):
@@ -91,7 +88,7 @@ def test_array_real_2d_F_scalar_add(language):
     f1(x1, a)
     f2(x2, a)
 
-    assert np.allclose(x1, x2, rtol=RTOL, atol=ATOL)
+    assert np.array_equal( x1, x2 )
 
 def test_array_real_2d_F_add(language):
 
@@ -105,7 +102,7 @@ def test_array_real_2d_F_add(language):
     f1(x1, a)
     f2(x2, a)
 
-    assert np.allclose(x1, x2, rtol=RTOL, atol=ATOL)
+    assert np.array_equal( x1, x2 )
 
 def test_array_int32_2d_F_complex_3d_expr(language):
 
@@ -133,7 +130,7 @@ def test_array_real_1d_complex_3d_expr(language):
     f1(x1, a)
     f2(x2, a)
 
-    assert np.allclose(x1, x2, rtol=RTOL, atol=ATOL)
+    assert np.array_equal( x1, x2 )
 
 def test_fib(language):
     f1 = python_annotations.fib
