### 2022-12-26 20d8e5b
issue_1254 - remove unnecessary approximate comparisons for floats
