+++ tests/epyccel/recognised_functions/test_numpy_funcs.py
@@ -4529,19 +4529,25 @@ def test_numpy_prod_array_like_1d(language):
     integer32 = randint(min_int32, max_int32, size = size, dtype=np.int32)
     integer64 = randint(min_int64, max_int64, size = size, dtype=np.int64)
 
-    fl = uniform(min_float / 2, max_float / 2, size = size)
-    fl32 = uniform(min_float32 / 2, max_float32 / 2, size = size)
-    fl32 = np.float32(fl32)
-    fl64 = uniform(min_float64 / 2, max_float64 / 2, size=size)
+    fl = uniform(-((-min_float) ** (1/5)), max_float ** (1/5), size = size)
 
-    cmplx128_from_float32 = uniform(low=-((-min_float32) ** (1/5)),
-                                    high=(max_float32 ** (1/5)), size = size) + \
-                            uniform(low=-((-min_float32) ** (1/5)),
-                                    high=(max_float32 ** (1/5)), size = size) * 1j
-    cmplx128_from_float64 = uniform(low=-((-min_float64) ** (1/5)),
-                                    high=(max_float64 ** (1/5)), size = size) + \
-                            uniform(low=-((-min_float64) ** (1/5)),
-                                    high=(max_float64 ** (1/5)), size = size) * 1j
+    min_ok_float32 = -((-min_float32) ** (1/5))
+    min_ok_float64 = -((-min_float64) ** (1/5))
+    max_ok_float32 = max_float32 ** (1/5)
+    max_ok_float64 = max_float64 ** (1/5)
+
+    fl32 = uniform(min_ok_float32, max_ok_float32, size = size)
+    fl32 = np.float32(fl32)
+    fl64 = uniform(min_ok_float64, max_ok_float64, size=size)
+
+    cmplx128_from_float32 = uniform(low=min_ok_float32/2,
+                                    high=max_ok_float32/2, size = size) + \
+                            uniform(low=min_ok_float32/2,
+                                    high=max_ok_float32/2, size = size) * 1j
+    cmplx128_from_float64 = uniform(low=min_ok_float64/2,
+                                    high=max_ok_float64/2, size = size) + \
+                            uniform(low=min_ok_float64/2,
+                                    high=max_ok_float64/2, size = size) * 1j
     # the result of the last operation is a Python complex type which has 8 bytes in the alignment,
     # that's why we need to convert it to a numpy.complex64 the needed type.
     cmplx64 = np.complex64(cmplx128_from_float32)
@@ -4560,6 +4566,17 @@ def test_numpy_prod_array_like_1d(language):
     assert np.isclose(epyccel_func(fl64), get_prod(fl64), rtol=RTOL, atol=ATOL)
     assert np.isclose(epyccel_func(cmplx64), get_prod(cmplx64), rtol=RTOL32, atol=ATOL32)
     assert np.isclose(epyccel_func(cmplx128), get_prod(cmplx128), rtol=RTOL, atol=ATOL)
+    assert matching_types(epyccel_func(bl), get_prod(bl))
+    assert matching_types(epyccel_func(integer8), get_prod(integer8))
+    assert matching_types(epyccel_func(integer16), get_prod(integer16))
+    assert matching_types(epyccel_func(integer), get_prod(integer))
+    assert matching_types(epyccel_func(integer32), get_prod(integer32))
+    assert matching_types(epyccel_func(integer64), get_prod(integer64))
+    assert matching_types(epyccel_func(fl), get_prod(fl))
+    assert matching_types(epyccel_func(fl32), get_prod(fl32))
+    assert matching_types(epyccel_func(fl64), get_prod(fl64))
[…]
