### 2022-12-26 4b7de8e
More explanation about 'efficient indexing'
