+++ developer_docs/order_docs.md
@@ -307,14 +307,12 @@ int main()
 
 ## Pyccel's Fortran code
 
-### Ordering in Fortran code
-
 As Fortran has arrays in the language there is no need to add special handling for arrays. Fortran ordered arrays (`order_f`) are already compatible with the Fortran language. They can therefore be passed to the function as they are.
 
 In order to pass C ordered arrays (`order_c`) and retain the shape and correct element placing to be compatible with Fortran, a transpose would be needed.
 In pyccel, we prefer to avoid unnecessary copies, so instead we pass the contiguous block of memory to Fortran and change how we index the array to ensure that we access the expected element.
 
-### Indexing arrays in Fortran
+### Ordering in Fortran code
 
 Fortran indexing does not occur in the same order as in C.
 If we take the following 2D array as an example:
@@ -336,122 +334,29 @@ for (int row = 0; row < 2; ++row) {
 ```
 
 while in Fortran we would do:
-
 ```Fortran
-# A.shape = (3,2)
-do column = 0, 2
-    do row = 0, 3
+# A.shape = (2,3)
+do column = 0, 3
+    do row = 0, 2
         A(row,column) = ....
     end do
 end do
 ```
 
-As you can see in the Fortran-ordered array the indices are inverted compared to how they would be printed in C.
-Pyccel therefore handles the printing of Fortran-ordered arrays by inverting the index order so as to preserve the most efficient indexing.
+As you can see in the Fortran-ordered array the indices are passed to the array in the same order, however the index does not point to the same location in memory.
+In C code the index `i_1, i_2, i_3` points to the element `i_1 * (n_2 * n_3) + i_2 * n_2 + i_3` in memory.
+In Fortran code the index `i_1, i_2, i_3` points to the element `i_1 + i_2 * n_1 + i_3 * (n_2 * n_3)` in memory.
 
-Example:
+### Order F
+Pyccel's translation of code with `order='F'` should look very similar to the original python code.
 
-```python
-import numpy as np
+Numpy's storage of the strides ensures that the first dimension is the contiguous dimension as in Fortran, so the code is equivalent for all element-wise operations.
 
-if __name__ == '__main__':
-    A = np.array([[1,2],[3,4],[5,6]], order='C')
-    B = np.array([[1,2],[3,4],[5,6]], order='F')
-    print(A.shape)  # (3,2)
-    print(B.shape)   # (3,2)
-    # print(A.ravel('K')) # array([1, 2, 3, 4, 5, 6])
-    # print(B.ravel('K')) # array([1, 3, 5, 2, 4, 6])
-
-    # Index optimally for A:
-    for row in range(3):
-        for column in range(2):
-            A[row,column] = ...
-
-    # Index optimally for B:
-    for column in range(2):
-        for row in range(3):
-            B[row,column] = ...
-```
-
-Will be translated to:
-
-```Fortran
-program prog_prog_tmp
-
-  use tmp
-
-  use, intrinsic :: ISO_C_Binding, only : i64 => C_INT64_T
-  implicit none
-
-  integer(i64), allocatable :: A(:,:)
-  integer(i64), allocatable :: B(:,:)
-  integer(i64) :: row
-  integer(i64) :: column
-
-  allocate(A(0:1_i64, 0:2_i64))
-  A = reshape([[1_i64, 2_i64], [3_i64, 4_i64], [5_i64, 6_i64]], [2_i64, &
-        3_i64])
[…]
