### 2022-12-26 e4e472d
Improve bounds
