+++ pyccel/codegen/compiling/compilers.py
@@ -300,11 +300,13 @@ class Compiler:
                 *j_code]
 
         with FileLock('.lock_acquisition.lock'):
+            print("Acquiring : ", os.path.abspath('.lock_acquisition.lock'))
             compile_obj.acquire_lock()
         try:
             self.run_command(cmd, verbose)
         finally:
             compile_obj.release_lock()
+            print("Releasing : ", os.path.abspath('.lock_acquisition.lock'))
 
     def compile_program(self, compile_obj, output_folder, verbose = False):
         """
@@ -340,11 +342,13 @@ class Compiler:
                 *libs_flags, *j_code]
 
         with FileLock('.lock_acquisition.lock'):
+            print("Acquiring : ", os.path.abspath('.lock_acquisition.lock'))
             compile_obj.acquire_lock()
         try:
             self.run_command(cmd, verbose)
         finally:
             compile_obj.release_lock()
+            print("Releasing : ", os.path.abspath('.lock_acquisition.lock'))
 
         return compile_obj.program_target
 
@@ -393,11 +397,13 @@ class Compiler:
                 '-o', file_out, *libs_flags]
 
         with FileLock('.lock_acquisition.lock'):
+            print("Acquiring : ", os.path.abspath('.lock_acquisition.lock'))
             compile_obj.acquire_lock()
         try:
             self.run_command(cmd, verbose)
         finally:
             compile_obj.release_lock()
+            print("Releasing : ", os.path.abspath('.lock_acquisition.lock'))
 
         return file_out
 
