### 2022-12-27 0fd03eb
Deal with PyObjects directly
