### 2022-12-27 8d9c4b9
Parse all arguments to a PyObject
