### 2022-12-27 e254fab
Change checks to deal with PyObjects directly
