### 2022-12-27 7b243ed
Simplify error printing
