### 2022-12-27 c576694
Remove unnecessary PyccelPyArrayObject type
