### 2022-12-27 4ee5f64
fix kernel launch configuration errors
