+++ pyccel/parser/semantic.py
@@ -121,7 +121,7 @@ from pyccel.ast.variable import DottedName, DottedVariable
 from pyccel.errors.errors import Errors
 from pyccel.errors.errors import PyccelSemanticError
 
-from pyccel.errors.messages import (PYCCEL_RESTRICTION_TODO, UNDERSCORE_NOT_A_THROWAWAY,
+from pyccel.errors.messages import (MISSING_KERNEL_CONFIGURATION,PYCCEL_RESTRICTION_TODO, UNDERSCORE_NOT_A_THROWAWAY,
         UNDEFINED_VARIABLE, IMPORTING_EXISTING_IDENTIFIED, INDEXED_TUPLE, LIST_OF_TUPLES,
         INVALID_INDICES, INCOMPATIBLE_ARGUMENT, INCOMPATIBLE_ORDERING,
         UNRECOGNISED_FUNCTION_CALL, STACK_ARRAY_SHAPE_UNPURE_FUNC, STACK_ARRAY_UNKNOWN_SHAPE,
@@ -807,7 +807,8 @@ class SemanticParser(BasicParser):
         =======
         new_expr : FunctionCall or PyccelInternalFunction
         """
-
+        if 'kernel' in func.decorators:
+            errors.report(MISSING_KERNEL_CONFIGURATION, symbol = expr, severity = 'fatal')
         if isinstance(func, PyccelFunctionDef):
             func = func.cls_name
             if func in (CudaThreadIdx, CudaBlockDim, CudaBlockIdx, CudaGridDim):
@@ -903,14 +904,27 @@ class SemanticParser(BasicParser):
                         severity='fatal')
             # TODO : type check the NUMBER OF BLOCKS 'numBlocks' and threads per block 'tpblock'
             if not isinstance(expr.numBlocks, LiteralInteger):
-                errors.report("Invalid Block number parameter for Kernel call",
+                try:
+                    numBlocks = self.scope.variables[expr.numBlocks]
+                    if not isinstance(numBlocks.dtype, NativeInteger):
+                        errors.report("Invalid Block number parameter for Kernel call",
+                        symbol = expr,
+                        severity='error')
+                except:
+                        errors.report("Invalid Block number parameter for Kernel call",
                         symbol = expr,
                         severity='error')
             if not isinstance(expr.tpblock, LiteralInteger):
-                errors.report("Invalid Thread per Block parameter for Kernel call",
+                try:
+                    tpblock = self.scope.variables[expr.tpblock]
+                    if not isinstance(tpblock.dtype, NativeInteger):
+                        errors.report("Invalid Thread per Block parameter for Kernel call",
                         symbol = expr,
                         severity='error')
-
+                except:
+                    errors.report("Invalid Thread per Block parameter for Kernel call",
+                        symbol = expr,
+                        severity='error')
             new_expr = KernelCall(func, args, expr.numBlocks, expr.tpblock, self._current_function)
 
             for a in new_expr.args:
