### 2022-12-27 b8c9394
Creation steps talk about np.array, adds spaces for new lines
