### 2022-12-27 e3a3db2
Don't raise error during type investigation (interfaces)
