+++ pyccel/stdlib/cwrapper_ndarrays/cwrapper_ndarrays.c
@@ -3,6 +3,7 @@
 /* or go to https://github.com/pyccel/pyccel/blob/master/LICENSE for full license details. */
 /* --------------------------------------------------------------------------------------- */
 
+#include <string.h>
 #include "cwrapper_ndarrays.h"
 
 /*
@@ -83,30 +84,30 @@ static npy_intp *_ndarray_to_numpy_shape(int64_t *nd_shape, int nd)
  *		a 	  : python array object
  *      dtype : desired data type enum
  * 	Returns		:
- *		return true if no error occurred otherwise it will return false
- *      and raise TypeError exception
+ *		return NULL if no error occurred otherwise it will return the
+ *      message to be reported in a TypeError exception
  * reference of the used c/python api function
  * -------------------------------------------
  * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_TYPE
  */
-bool	check_pyarray_dtype(PyArrayObject *a, int dtype)
+static char*	_check_pyarray_dtype(PyArrayObject *a, int dtype)
 {
 	int current_dtype;
 
 	if (dtype == NO_TYPE_CHECK)
-		return true;
+		return NULL;
 
 	current_dtype = PyArray_TYPE(a);
 	if (current_dtype != dtype)
 	{
-		PyErr_Format(PyExc_TypeError,
-			"argument dtype must be %s, not %s",
+        char* error = (char *)malloc(200);
+        sprintf(error, "argument dtype must be %s, not %s",
 			dataTypes[dtype],
 			dataTypes[current_dtype]);
-		return false;
+		return error;
 	}
 
-	return true;
+	return NULL;
 }
 
 /*
@@ -118,26 +119,27 @@ bool	check_pyarray_dtype(PyArrayObject *a, int dtype)
  *		a 	  : python array object
  *      rank  : desired rank
  * 	Returns		:
- *		return true if no error occurred otherwise it will return false
- *      and raise TypeError exception
+ *		return NULL if no error occurred otherwise it will return the
+ *      message to be reported in a TypeError exception
  * reference of the used c/python api function
  * -------------------------------------------
  * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_NDIM
  */
-static bool _check_pyarray_rank(PyArrayObject *a, int rank)
+static char* _check_pyarray_rank(PyArrayObject *a, int rank)
 {
 	int	current_rank;
 
 	current_rank = PyArray_NDIM(a);
 	if (current_rank != rank)
 	{
-		PyErr_Format(PyExc_TypeError, "argument rank must be %d, not %d",
+        char* error = (char *)malloc(200);
+        sprintf(error, "argument rank must be %d, not %d",
 			rank,
 			current_rank);
-		return false;
+		return error;
 	}
 
-	return true;
+	return NULL;
 }
 
 /*
@@ -149,27 +151,27 @@ static bool _check_pyarray_rank(PyArrayObject *a, int rank)
  *		a 	  : python array object
  *      flag  : desired order
  * 	Returns		:
- *		return true if no error occurred otherwise it will return false
- *      and raise NotImplementedError exception
+ *		return NULL if no error occurred otherwise it will return the
+ *      message to be reported in a TypeError exception
  * reference of the used c/python api function
  * -------------------------------------------
  * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_CHKFLAGS
  */
-static bool _check_pyarray_order(PyArrayObject *a, int flag)
+static char* _check_pyarray_order(PyArrayObject *a, int flag)
 {
 
[…]
