+++ pyccel/stdlib/cwrapper_ndarrays/cwrapper_ndarrays.h
@@ -45,6 +45,7 @@ PyObject* fortran_ndarray_to_pyarray(t_ndarray *o);
 
 /* arrays checkers and helpers */
 bool	pyarray_check(PyObject *o, int dtype, int rank, int flag);
+bool	is_numpy_array(PyObject *o, int dtype, int rank, int flag);
 
 void    *nd_data(t_ndarray *a);
 int     nd_ndim(t_ndarray *a, int n);
