+++ pyccel/ast/numpy_wrapper.py
@@ -20,8 +20,6 @@ from .internals         import get_final_precision
 
 from .literals          import LiteralInteger
 
-from .operators         import PyccelEq
-
 from .variable          import Variable
 
 from ..errors.errors   import Errors
