+++ pyccel/ast/numpyext.py
@@ -12,7 +12,7 @@ import operator
 
 import numpy
 
-from pyccel.errors.errors import Errors, PyccelError
+from pyccel.errors.errors import Errors
 from pyccel.errors.messages import WRONG_LINSPACE_ENDPOINT, NON_LITERAL_KEEP_DIMS, NON_LITERAL_AXIS
 
 from pyccel.utilities.stage import PyccelStage
