### 2022-12-27 db709c9
Avoid checking for attribute 'decorators' for Interface objects
