### 2022-12-28 de588d6
improves bad if statment
