### 2022-12-28 1f2e3e6
better if statement, fixes print_ndarray_memory bug
