### 2022-12-28 990a791
adds missing parenthesis, uses ObjectAddress instead of '&' hardcode
