### 2022-12-28 3a5902b
Uses ObjectAddress, and calculates it once
