### 2022-12-29 b27fcab
removed duplicated const printer
