### 2022-12-29 17bec94
Using Declare and Allocate instead of hardcoding
