### 2022-12-29 7969053
use self._print(lhs) instead of lhs_address
