### 2022-12-29 2a57e4a
Use non-fatal errors where possible
