### 2022-12-29 c0bd01d
Add ScopedNode to __all__
