### 2022-12-29 4b83a64
Add docs for the syntactic stage
