+++ pyccel/codegen/compiling/basic.py
@@ -220,7 +220,6 @@ class CompileObj:
         """
         if self.has_target_file:
             self._lock.acquire()
-            print("Acquired ", self._lock)
 
     def release_lock(self):
         """
@@ -236,7 +235,6 @@ class CompileObj:
         """
         if self.has_target_file:
             self._lock.release()
-            print("Released", self._lock)
 
     @property
     def accelerators(self):
