### 2022-12-29 f056812
Simplify name definition (handles renaming)
