### 2022-12-29 b694f61
Add missing locks
