### 2022-12-29 2802443
Add debugging info
