### 2022-12-29 38dcc05
Run relevant tests
