### 2022-12-29 cb5c4cc
Split test to_host and to device
