### 2022-12-29 1333089
Remove useless touch
