### 2022-12-30 1f20034
Remove unnecessary locks. Don't lock lock file
