### 2022-12-30 4859bc9
Run array tests together first
