### 2022-12-30 704af12
Add error tests
