+++ tests/errors/syntax_errors/nargs.py
@@ -1,0 +1,3 @@
+
+def f(*args):
+    pass
