+++ tests/errors/syntax_errors/not_implemented.py
@@ -1,0 +1,4 @@
+
+a @= b
+
+c = all(True for i in range(5))
