+++ tests/errors/syntax_errors/not_implemented.py
@@ -1,3 +1,6 @@
+# pylint: disable=missing-function-docstring, missing-module-docstring/
+
+b = 1
 
 a @= b
 
