+++ tests/errors/syntax_errors/nargs.py
@@ -1,3 +1,4 @@
+# pylint: disable=missing-function-docstring, missing-module-docstring/
 
 def f(*args):
     pass
