### 2022-12-30 f6655a4
EmptyNodes not needed
