### 2022-12-30 d85f824
set new_expr to AugAssign of Variable is pointer
