### 2022-12-30 fb39e98
Add examples. Explain SyntaxParser._in_lhs_assign
