+++ developer_docs/syntactic_stage.md
@@ -3,7 +3,7 @@
 The syntactic stage is described by the file [pyccel.parser.syntactic](../pyccel/parser/syntactic.py)
 
 The syntactic stage serves 4 main purposes:
-1.  [**Navigation and AST Creation**](#Navigation-and-ast-creation) : Convert Python's [AST](https://docs.python.org/3/library/ast.html) (abstract syntax tree) representation of the python file to Pyccel's AST representation (objects of the classes in the folder [pyccel.ast](../pyccel/ast))
+1.  [**Navigation and AST Creation**](#Navigation-and-ast-creation) : Convert Python's [AST](https://docs.Python.org/3/library/ast.html) (abstract syntax tree) representation of the Python file to Pyccel's AST representation (objects of the classes in the folder [pyccel.ast](../pyccel/ast))
 2.  [**Errors**](#Errors) : Raise an error for any syntax used that is not yet supported by pyccel
 3.  [**Headers**](#Headers) : Convert header comments from strings to Pyccel's AST representation
 4.  [**Scoping**](#Scoping) : Collect the name of all variables in each scope (see [scope](scope.md) for more details) to ensure no name collisions can occur if pyccel generates Variable names
@@ -24,6 +24,36 @@ def _visit_ClassName(self, stmt):
     ...
 ```
 Each of these functions should internally call the `_visit` function on each of the elements of the object to obtain Pyccel AST nodes which can be combined to create a Pyccel AST node representing the current object.
+
+### Example
+
+Consider for example the Python expression:
+```python
+a += b
+```
+The Python ast module parses this as:
+```python
+AugAssign(target=Name(id='a', ctx=Store()), op=Add(), value=Name(id='b', ctx=Load()))
+```
+and we want to save this information into Pyccel's [`pyccel.ast.core.AugAssign`](../pyccel/ast/core.py).
+We therefore need to visit the 2 members of `ast.AugAssign`:
+- target
+- value
+and use the third (`op`) to correctly construct the `pyccel.ast.core.AugAssign`.
+
+The final code therefore ressembles the following:
+```python
+def _visit_AugAssign(self, stmt):
+    lhs = self._visit(stmt.target)
+    rhs = self._visit(stmt.value)
+    if isinstance(stmt.op, ast.Add):
+        return AugAssign(lhs, '+', rhs)
+    elif ...:
+        ...
+    else:
+        return errors.report(PYCCEL_RESTRICTION_TODO, symbol = stmt,
+                severity='error')
+```
 
 ## Errors
 
@@ -51,10 +81,35 @@ These instances can then be inserted into the abstract syntax tree.
 
 ## Scoping
 
-The final purpose of the syntactic stage is to collect all names used in the code.
+The final purpose of the syntactic stage is to collect all names declared in the code.
 This is important to avoid name collisions if pyccel creates temporaries or requires additional names.
 The names are saved in the scope (for more details see [scope](scope.md)).
-Whenever a symbol is encountered it should be saved to the scope using the function `self.scope.insert_symbol`.
+Whenever a symbol is encountered in a declaration context, it should be saved to the scope using the function `self.scope.insert_symbol`.
+This is usually done in the `_visit_Name` function, however this function is not aware of the context so it cannot determine whether it is a declaration.
+To get round this the class has the property `SyntaxParser._in_lhs_assign`, which should be `True` in a declaration context, and false elsewhere.
+
+Consider for example a for loop. Such a loop has 3 main parts (which are each members of `ast.For`):
+-  target  (`ast.For.target`)
+-  iterable  (`ast.For.iter`)
+-  body (`ast.For.body`)
+
+such that a for loop is defined as:
+```python
+for target in iterable:
+   body
+```
+
[…]
