+++ developer_docs/overview.md
@@ -1,16 +1,22 @@
 ## Developer Setup
 
-Before beginning any development in pyccel, it is important to ensure pyccel is correctly installed **from source in development mode** as described [here](../README.md#from-sources). If this step is not followed then any changes made to source will not be used when `pyccel` or `epyccel` are used.
+Before beginning any development in Pyccel, it is important to ensure Pyccel is correctly installed **from source in development mode** as described [here](../README.md#from-sources). If this step is not followed then any changes made to source will not be used when `Pyccel` or `ePyccel` are used.
 
 ## Overview
 
 Pyccel's development is split into 4 main stages:
 
+- [Syntactic Stage](#Syntactic-stage) (for more details see [syntactic stage](syntactic_stage.md))
+- [Semantic Stage](#Semantic-stage)
+- [Code Generation Stage](#Code-generation-stage)
+- [Wrapping Stage](#Wrapping-stage)
+- [Compliation Stage](#Compilation-stage)
+
 ### Syntactic Stage
 
-Pyccel uses Python's [ast module](https://docs.python.org/3/library/ast.html) to read the input file(s). The ast does not store information in the same way as the rest of Pyccel so this stage exists to **convert Python's ast to Pyccel's ast**. The related code can be found in [parser/syntactic.py](../pyccel/parser/syntactic.py).
+Pyccel uses Python's [ast module](https://docs.python.org/3/library/ast.html) to read the input file(s). The ast does not store information in the same way as the rest of Pyccel so this stage exists to **convert Python's ast to Pyccel's ast**. The related code can be found in [parser/syntactic.py](../Pyccel/parser/syntactic.py).
 
-The syntactic stage also handles parsing header comments. This is managed using [textx](http://textx.github.io/textX/stable/). The files describing the _textx_ grammar are found in the folder [parser/grammar](../pyccel/parser/grammar). From these files _textx_ generates instances of the classes found in the folder [parser/syntax](../pyccel/parser/syntax).
+The syntactic stage also handles parsing header comments. This is managed using [textx](http://textx.github.io/textX/stable/). The files describing the _textx_ grammar are found in the folder [parser/grammar](../Pyccel/parser/grammar). From these files _textx_ generates instances of the classes found in the folder [parser/syntax](../Pyccel/parser/syntax).
 
 #### Advanced Comments
 
@@ -18,23 +24,27 @@ The role of this stage has decreased significantly since we moved from [redbaron
 
 ### Semantic Stage
 
-This is the most important stage in pyccel. It is here that all the information about types is calculated. This stage strives to be **language-agnostic**; this means for example, that additional variables required to handle problems appearing in one specific language should not be created here.
+This is the most important stage in Pyccel. It is here that all the information about types is calculated. This stage strives to be **language-agnostic**; this means for example, that additional variables required to handle problems appearing in one specific language should not be created here.
 
-When adding functions to this stage the aim is often to create a `PyccelAstNode` (see [ast/basic.py](../pyccel/ast/basic.py)) and correctly define all of its parameters. This information is sometimes readily available (e.g. the type of a `PyccelAdd` can be derived from the type of the variables passed to it), but sometimes the information must be collected from elsewhere (e.g. when creating a `Variable` from a `PyccelSymbol` (roughly equivalent to a string). In this case information is needed from a `Scope` instance which is stored in the `scope`.
[…]
