+++ developer_docs/overview.md
@@ -6,11 +6,11 @@ Before beginning any development in Pyccel, it is important to ensure Pyccel is 
 
 Pyccel's development is split into 4 main stages:
 
-- [Syntactic Stage](#Syntactic-stage) (for more details see [syntactic stage](syntactic_stage.md))
-- [Semantic Stage](#Semantic-stage)
-- [Code Generation Stage](#Code-generation-stage)
-- [Wrapping Stage](#Wrapping-stage)
-- [Compliation Stage](#Compilation-stage)
+-   [Syntactic Stage](#Syntactic-stage) (for more details see [syntactic stage](syntactic_stage.md))
+-   [Semantic Stage](#Semantic-stage)
+-   [Code Generation Stage](#Code-generation-stage)
+-   [Wrapping Stage](#Wrapping-stage)
+-   [Compliation Stage](#Compilation-stage)
 
 ### Syntactic Stage
 
