### 2022-12-30 280bf39
prevent dealocation of alias variables
