### 2022-12-30 d607b56
Clarify sentence
