### 2022-12-30 aa4e20a
prevent pointer variables to be reassign in case if augassign
