+++ docs/source/pyccel.parser.rst
@@ -1,0 +1,77 @@
+pyccel.parser package
+=====================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyccel.parser.syntax
+
+Submodules
+----------
+
+pyccel.parser.base module
+-------------------------
+
+.. automodule:: pyccel.parser.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.extend\_tree module
+---------------------------------
+
+.. automodule:: pyccel.parser.extend_tree
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.parser module
+---------------------------
+
+.. automodule:: pyccel.parser.parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.scope module
+--------------------------
+
+.. automodule:: pyccel.parser.scope
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.semantic module
+-----------------------------
+
+.. automodule:: pyccel.parser.semantic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.syntactic module
+------------------------------
+
+.. automodule:: pyccel.parser.syntactic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.parser.utilities module
+------------------------------
+
+.. automodule:: pyccel.parser.utilities
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
