+++ docs/source/pyccel.commands.rst
@@ -1,0 +1,37 @@
+pyccel.commands package
+=======================
+
+Submodules
+----------
+
+pyccel.commands.console module
+------------------------------
+
+.. automodule:: pyccel.commands.console
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.commands.pyccel\_clean module
+------------------------------------
+
+.. automodule:: pyccel.commands.pyccel_clean
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.commands.pyccel\_init module
+-----------------------------------
+
+.. automodule:: pyccel.commands.pyccel_init
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.commands
+   :members:
+   :undoc-members:
+   :show-inheritance:
