+++ docs/source/pyccel.stdlib.rst
@@ -1,0 +1,20 @@
+pyccel.stdlib package
+=====================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyccel.stdlib.external
+   pyccel.stdlib.internal
+   pyccel.stdlib.parallel
+
+Module contents
+---------------
+
+.. automodule:: pyccel.stdlib
+   :members:
+   :undoc-members:
+   :show-inheritance:
