+++ docs/source/pyccel.codegen.rst
@@ -1,0 +1,54 @@
+pyccel.codegen package
+======================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyccel.codegen.compiling
+   pyccel.codegen.printing
+
+Submodules
+----------
+
+pyccel.codegen.codegen module
+-----------------------------
+
+.. automodule:: pyccel.codegen.codegen
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.codegen.pipeline module
+------------------------------
+
+.. automodule:: pyccel.codegen.pipeline
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.codegen.python\_wrapper module
+-------------------------------------
+
+.. automodule:: pyccel.codegen.python_wrapper
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.codegen.utilities module
+-------------------------------
+
+.. automodule:: pyccel.codegen.utilities
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.codegen
+   :members:
+   :undoc-members:
+   :show-inheritance:
