+++ docs/source/pyccel.utilities.rst
@@ -1,0 +1,37 @@
+pyccel.utilities package
+========================
+
+Submodules
+----------
+
+pyccel.utilities.metaclasses module
+-----------------------------------
+
+.. automodule:: pyccel.utilities.metaclasses
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.utilities.stage module
+-----------------------------
+
+.. automodule:: pyccel.utilities.stage
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.utilities.strings module
+-------------------------------
+
+.. automodule:: pyccel.utilities.strings
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.utilities
+   :members:
+   :undoc-members:
+   :show-inheritance:
