+++ docs/source/pyccel.complexity.rst
@@ -1,0 +1,37 @@
+pyccel.complexity package
+=========================
+
+Submodules
+----------
+
+pyccel.complexity.arithmetic module
+-----------------------------------
+
+.. automodule:: pyccel.complexity.arithmetic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.complexity.basic module
+------------------------------
+
+.. automodule:: pyccel.complexity.basic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.complexity.memory module
+-------------------------------
+
+.. automodule:: pyccel.complexity.memory
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.complexity
+   :members:
+   :undoc-members:
+   :show-inheritance:
