+++ docs/source/modules.rst
@@ -1,0 +1,10 @@
+pyccel-master
+=============
+
+.. toctree::
+   :maxdepth: 4
+
+   pyccel
+   setup
+   test
+   test
