+++ docs/source/pyccel.errors.rst
@@ -1,0 +1,29 @@
+pyccel.errors package
+=====================
+
+Submodules
+----------
+
+pyccel.errors.errors module
+---------------------------
+
+.. automodule:: pyccel.errors.errors
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.errors.messages module
+-----------------------------
+
+.. automodule:: pyccel.errors.messages
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.errors
+   :members:
+   :undoc-members:
+   :show-inheritance:
