+++ docs/source/pyccel.stdlib.parallel.rst
@@ -1,0 +1,21 @@
+pyccel.stdlib.parallel package
+==============================
+
+Submodules
+----------
+
+pyccel.stdlib.parallel.mpi module
+---------------------------------
+
+.. automodule:: pyccel.stdlib.parallel.mpi
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.stdlib.parallel
+   :members:
+   :undoc-members:
+   :show-inheritance:
