+++ docs/source/pyccel.codegen.compiling.rst
@@ -1,0 +1,29 @@
+pyccel.codegen.compiling package
+================================
+
+Submodules
+----------
+
+pyccel.codegen.compiling.basic module
+-------------------------------------
+
+.. automodule:: pyccel.codegen.compiling.basic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyccel.codegen.compiling.compilers module
+-----------------------------------------
+
+.. automodule:: pyccel.codegen.compiling.compilers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyccel.codegen.compiling
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
