### 2022-12-30 2a94f14
Remove slot from CupyRavel
Co-authored-by: EmilyBourne <louise.bourne@gmail.com>
