### 2022-12-30 42c9942
Only Apply NativeVoid check on PyccelSymbol
