### 2022-12-30 7e71c89
add test for invalid kernel configuration
