### 2022-12-31 4905573
Add docstring for memory_location and current_context
