### 2022-12-31 c427bb5
Use f-string in copy_CudaArray_Data
