### 2022-12-31 b4e5670
Disable missing docstring and add missing final new line is ccuda tests
