### 2022-12-31 de6dc5c
Use f-string in memcpy_kind
