### 2023-01-01 d83f526
Fix codacy problems
