### 2023-01-01 bb1dbba
Try to fix django not configured pylint error
