### 2023-01-01 72377e1
update test to return first value
