+++ tests/epyccel/test_epyccel_augassign.py
@@ -16,23 +16,13 @@ def test_augassign_add_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones(5, dtype=int)
-    x1_float   = np.ones(5, dtype=float)
-    x1_complex = np.ones(5, dtype=complex)
-    x2_int     = np.ones(5, dtype=int)
-    x2_float   = np.ones(5, dtype=float)
-    x2_complex = np.ones(5, dtype=complex)
+    x_int     = np.ones(5, dtype=int)
+    x_float   = np.ones(5, dtype=float)
+    x_complex = np.ones(5, dtype=complex)
 
-    f_int(x1_int)
-    f_float(x1_float)
-    f_complex(x1_complex)
-    f_int_epyc(x2_int)
-    f_float_epyc(x2_float)
-    f_complex_epyc(x2_complex)
-
-    assert x1_int == x2_int and x1_int.dtype == x2_int.dtype
-    assert x1_float == x2_float and x1_float.dtype == x2_float.dtype
-    assert x1_complex == x2_complex and x1_complex.dtype == x2_complex.dtype
+    assert f_int(x_int) == f_int_epyc(x_int) == 42
+    assert f_float(x_float) == f_float_epyc(x_float) == 4.2
+    assert f_complex(x_complex) == f_complex_epyc(x_complex) == (4.0 + 2.0j)
 
 def test_augassign_add_2d(language):
     f_int     = mod.augassign_add_2d_int
@@ -42,23 +32,13 @@ def test_augassign_add_2d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones((5, 5), dtype=int)
-    x1_float   = np.ones((5, 5), dtype=float)
-    x1_complex = np.ones((5, 5), dtype=complex)
-    x2_int     = np.ones((5, 5), dtype=int)
-    x2_float   = np.ones((5, 5), dtype=float)
-    x2_complex = np.ones((5, 5), dtype=complex)
+    x_int     = np.ones((5, 5), dtype=int)
+    x_float   = np.ones((5, 5), dtype=float)
+    x_complex = np.ones((5, 5), dtype=complex)
 
-    f_int(x1_int)
-    f_float(x1_float)
-    f_complex(x1_complex)
-    f_int_epyc(x2_int)
-    f_float_epyc(x2_float)
-    f_complex_epyc(x2_complex)
-
-    assert x1_int == x2_int and x1_int.dtype == x2_int.dtype
-    assert x1_float == x2_float and x1_float.dtype == x2_float.dtype
-    assert x1_complex == x2_complex and x1_complex.dtype == x2_complex.dtype
+    assert f_int(x_int) == f_int_epyc(x_int) == 42
+    assert f_float(x_float) == f_float_epyc(x_float) == 4.2
+    assert f_complex(x_complex) == f_complex_epyc(x_complex) == (4.0 + 2.0j)
 
 
 # -= tests
@@ -71,23 +51,13 @@ def test_augassign_sub_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones(5, dtype=int)
-    x1_float   = np.ones(5, dtype=float)
-    x1_complex = np.ones(5, dtype=complex)
-    x2_int     = np.ones(5, dtype=int)
-    x2_float   = np.ones(5, dtype=float)
-    x2_complex = np.ones(5, dtype=complex)
[…]
