### 2023-01-01 41fa424
Remove disable=django-not-configure
