+++ tests/epyccel/modules/augassign.py
@@ -6,42 +6,36 @@ from pyccel.decorators import types
 
 @types('int[:]')
 def augassign_add_1d_int(a):
-    import numpy as np
     b = a
     b += 42
     return b[0]
 
 @types('float[:]')
 def augassign_add_1d_float(a):
-    import numpy as np
     b = a
     b += 4.2
     return b[0]
 
 @types('complex[:]')
 def augassign_add_1d_complex(a):
-    import numpy as np
     b = a
     b += (4.0 + 2.0j)
     return b[0]
 
 @types('int[:,:]')
 def augassign_add_2d_int(a):
-    import numpy as np
     b = a
     b += 42
     return b[0][0]
 
 @types('float[:,:]')
 def augassign_add_2d_float(a):
-    import numpy as np
     b = a
     b += 4.2
     return b[0][0]
 
 @types('complex[:,:]')
 def augassign_add_2d_complex(a):
-    import numpy as np
     b = a
     b += (4.0 + 2.0j)
     return b[0][0]
@@ -50,42 +44,36 @@ def augassign_add_2d_complex(a):
 
 @types('int[:]')
 def augassign_sub_1d_int(a):
-    import numpy as np
     b = a
     b -= 42
     return b[0]
 
 @types('float[:]')
 def augassign_sub_1d_float(a):
-    import numpy as np
     b = a
     b -= 4.2
     return b[0]
 
 @types('complex[:]')
 def augassign_sub_1d_complex(a):
-    import numpy as np
     b = a
     b -= (4.0 + 2.0j)
     return b[0]
 
 @types('int[:,:]')
 def augassign_sub_2d_int(a):
-    import numpy as np
     b = a
     b -= 42
     return b[0][0]
 
 @types('float[:,:]')
 def augassign_sub_2d_float(a):
-    import numpy as np
     b = a
     b -= 4.2
     return b[0][0]
 
 @types('complex[:,:]')
 def augassign_sub_2d_complex(a):
-    import numpy as np
     b = a
     b -= (4.0 + 2.0j)
     return b[0][0]
@@ -94,42 +82,36 @@ def augassign_sub_2d_complex(a):
 
 @types('int[:]')
 def augassign_mul_1d_int(a):
-    import numpy as np
     b = a
     b *= 42
     return b[0]
 
 @types('float[:]')
 def augassign_mul_1d_float(a):
-    import numpy as np
     b = a
     b *= 4.2
     return b[0]
 
 @types('complex[:]')
 def augassign_mul_1d_complex(a):
-    import numpy as np
     b = a
     b *= (4.0 + 2.0j)
     return b[0]
 
 @types('int[:,:]')
 def augassign_mul_2d_int(a):
-    import numpy as np
     b = a
     b *= 42
     return b[0][0]
 
 @types('float[:,:]')
 def augassign_mul_2d_float(a):
-    import numpy as np
     b = a
     b *= 4.2
     return b[0][0]
 
 @types('complex[:,:]')
 def augassign_mul_2d_complex(a):
-    import numpy as np
     b = a
     b *= (4.0 + 2.0j)
     return b[0][0]
@@ -138,42 +120,36 @@ def augassign_mul_2d_complex(a):
 
 @types('int[:]')
 def augassign_div_1d_int(a):
-    import numpy as np
     b = a
     b /= 42
     return b[0]
 
 @types('float[:]')
 def augassign_div_1d_float(a):
-    import numpy as np
     b = a
     b /= 4.2
     return b[0]
 
 @types('complex[:]')
 def augassign_div_1d_complex(a):
-    import numpy as np
     b = a
     b /= (4.0 + 2.0j)
     return b[0]
 
 @types('int[:,:]')
[…]
