+++ tests/epyccel/test_epyccel_augassign.py
@@ -16,13 +16,16 @@ def test_augassign_add_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x_int     = np.ones(5, dtype=int)
-    x_float   = np.ones(5, dtype=float)
-    x_complex = np.ones(5, dtype=complex)
+    x1_int     = np.ones(5, dtype=int)
+    x1_float   = np.ones(5, dtype=float)
+    x1_complex = np.ones(5, dtype=complex)
+    x2_int     = np.ones(5, dtype=int)
+    x2_float   = np.ones(5, dtype=float)
+    x2_complex = np.ones(5, dtype=complex)
 
-    assert f_int(x_int) == f_int_epyc(x_int) == 42
-    assert f_float(x_float) == f_float_epyc(x_float) == 4.2
-    assert f_complex(x_complex) == f_complex_epyc(x_complex) == (4.0 + 2.0j)
+    assert f_int(x1_int) == f_int_epyc(x2_int) == 42
+    assert f_float(x1_float) == f_float_epyc(x2_float) == 4.2
+    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) == (4.0 + 2.0j)
 
 def test_augassign_add_2d(language):
     f_int     = mod.augassign_add_2d_int
@@ -32,13 +35,16 @@ def test_augassign_add_2d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x_int     = np.ones((5, 5), dtype=int)
-    x_float   = np.ones((5, 5), dtype=float)
-    x_complex = np.ones((5, 5), dtype=complex)
+    x1_int     = np.ones((5, 5), dtype=int)
+    x1_float   = np.ones((5, 5), dtype=float)
+    x1_complex = np.ones((5, 5), dtype=complex)
+    x2_int     = np.ones((5, 5), dtype=int)
+    x2_float   = np.ones((5, 5), dtype=float)
+    x2_complex = np.ones((5, 5), dtype=complex)
 
-    assert f_int(x_int) == f_int_epyc(x_int) == 42
-    assert f_float(x_float) == f_float_epyc(x_float) == 4.2
-    assert f_complex(x_complex) == f_complex_epyc(x_complex) == (4.0 + 2.0j)
+    assert f_int(x1_int) == f_int_epyc(x2_int) == 42
+    assert f_float(x1_float) == f_float_epyc(x2_float) == 4.2
+    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) == (4.0 + 2.0j)
 
 
 # -= tests
@@ -51,13 +57,16 @@ def test_augassign_sub_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x_int     = np.ones(5, dtype=int)
-    x_float   = np.ones(5, dtype=float)
-    x_complex = np.ones(5, dtype=complex)
+    x1_int     = np.ones(5, dtype=int)
+    x1_float   = np.ones(5, dtype=float)
+    x1_complex = np.ones(5, dtype=complex)
+    x2_int     = np.ones(5, dtype=int)
+    x2_float   = np.ones(5, dtype=float)
+    x2_complex = np.ones(5, dtype=complex)
 
-    assert f_int(x_int) == f_int_epyc(x_int) == 42
-    assert f_float(x_float) == f_float_epyc(x_float) == 4.2
[…]
