### 2023-01-01 005c8c5
replaces np.ones with np.zeros
