+++ tests/epyccel/test_epyccel_augassign.py
@@ -16,16 +16,16 @@ def test_augassign_add_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones(5, dtype=int)
-    x1_float   = np.ones(5, dtype=float)
-    x1_complex = np.ones(5, dtype=complex)
-    x2_int     = np.ones(5, dtype=int)
-    x2_float   = np.ones(5, dtype=float)
-    x2_complex = np.ones(5, dtype=complex)
+    x1_int     = np.zeros(5, dtype=int)
+    x1_float   = np.zeros(5, dtype=float)
+    x1_complex = np.zeros(5, dtype=complex)
+    x2_int     = np.zeros(5, dtype=int)
+    x2_float   = np.zeros(5, dtype=float)
+    x2_complex = np.zeros(5, dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) == 42
-    assert f_float(x1_float) == f_float_epyc(x2_float) == 4.2
-    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) == (4.0 + 2.0j)
+    assert f_int(x1_int) == f_int_epyc(x2_int)
+    assert f_float(x1_float) == f_float_epyc(x2_float)
+    assert f_complex(x1_complex) == f_complex_epyc(x2_complex)
 
 def test_augassign_add_2d(language):
     f_int     = mod.augassign_add_2d_int
@@ -35,16 +35,16 @@ def test_augassign_add_2d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones((5, 5), dtype=int)
-    x1_float   = np.ones((5, 5), dtype=float)
-    x1_complex = np.ones((5, 5), dtype=complex)
-    x2_int     = np.ones((5, 5), dtype=int)
-    x2_float   = np.ones((5, 5), dtype=float)
-    x2_complex = np.ones((5, 5), dtype=complex)
+    x1_int     = np.zeros((5, 5), dtype=int)
+    x1_float   = np.zeros((5, 5), dtype=float)
+    x1_complex = np.zeros((5, 5), dtype=complex)
+    x2_int     = np.zeros((5, 5), dtype=int)
+    x2_float   = np.zeros((5, 5), dtype=float)
+    x2_complex = np.zeros((5, 5), dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) == 42
-    assert f_float(x1_float) == f_float_epyc(x2_float) == 4.2
-    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) == (4.0 + 2.0j)
+    assert f_int(x1_int) == f_int_epyc(x2_int)
+    assert f_float(x1_float) == f_float_epyc(x2_float)
+    assert f_complex(x1_complex) == f_complex_epyc(x2_complex)
 
 
 # -= tests
@@ -57,16 +57,16 @@ def test_augassign_sub_1d(language):
     f_float_epyc   = epyccel(f_float, language = language)
     f_complex_epyc = epyccel(f_complex, language = language)
 
-    x1_int     = np.ones(5, dtype=int)
-    x1_float   = np.ones(5, dtype=float)
-    x1_complex = np.ones(5, dtype=complex)
-    x2_int     = np.ones(5, dtype=int)
-    x2_float   = np.ones(5, dtype=float)
-    x2_complex = np.ones(5, dtype=complex)
+    x1_int     = np.zeros(5, dtype=int)
[…]
