### 2023-01-01 f903e27
added all_equal check and removed true_div for int
