### 2023-01-01 699b031
replaces all_equal with array_equal
