+++ tests/epyccel/test_epyccel_augassign.py
@@ -23,9 +23,16 @@ def test_augassign_add_1d(language):
     x2_float   = np.zeros(5, dtype=float)
     x2_complex = np.zeros(5, dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) and np.all_equal(x1_int, x2_int)
-    assert f_float(x1_float) == f_float_epyc(x2_float) and np.all_equal(x1_float, x2_float)
-    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) and np.all_equal(x1_complex, x2_complex)
+    f_int(x1_int)
+    f_float(x1_float)
+    f_complex(x1_complex)
+    f_int_epyc(x2_int)
+    f_float_epyc(x2_float)
+    f_complex_epyc(x2_complex)
+
+    assert np.array_equal(x1_int, x2_int)
+    assert np.array_equal(x1_float, x2_float)
+    assert np.array_equal(x1_complex, x2_complex)
 
 def test_augassign_add_2d(language):
     f_int     = mod.augassign_add_2d_int
@@ -42,9 +49,16 @@ def test_augassign_add_2d(language):
     x2_float   = np.zeros((5, 5), dtype=float)
     x2_complex = np.zeros((5, 5), dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) and np.all_equal(x1_int, x2_int)
-    assert f_float(x1_float) == f_float_epyc(x2_float) and np.all_equal(x1_float, x2_float)
-    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) and np.all_equal(x1_complex, x2_complex)
+    f_int(x1_int)
+    f_float(x1_float)
+    f_complex(x1_complex)
+    f_int_epyc(x2_int)
+    f_float_epyc(x2_float)
+    f_complex_epyc(x2_complex)
+
+    assert np.array_equal(x1_int, x2_int)
+    assert np.array_equal(x1_float, x2_float)
+    assert np.array_equal(x1_complex, x2_complex)
 
 
 # -= tests
@@ -64,9 +78,16 @@ def test_augassign_sub_1d(language):
     x2_float   = np.zeros(5, dtype=float)
     x2_complex = np.zeros(5, dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) and np.all_equal(x1_int, x2_int)
-    assert f_float(x1_float) == f_float_epyc(x2_float) and np.all_equal(x1_float, x2_float)
-    assert f_complex(x1_complex) == f_complex_epyc(x2_complex) and np.all_equal(x1_complex, x2_complex)
+    f_int(x1_int)
+    f_float(x1_float)
+    f_complex(x1_complex)
+    f_int_epyc(x2_int)
+    f_float_epyc(x2_float)
+    f_complex_epyc(x2_complex)
+
+    assert np.array_equal(x1_int, x2_int)
+    assert np.array_equal(x1_float, x2_float)
+    assert np.array_equal(x1_complex, x2_complex)
 
 def test_augassign_sub_2d(language):
     f_int     = mod.augassign_sub_2d_int
@@ -83,9 +104,16 @@ def test_augassign_sub_2d(language):
     x2_float   = np.zeros((5, 5), dtype=float)
     x2_complex = np.zeros((5, 5), dtype=complex)
 
-    assert f_int(x1_int) == f_int_epyc(x2_int) and np.all_equal(x1_int, x2_int)
-    assert f_float(x1_float) == f_float_epyc(x2_float) and np.all_equal(x1_float, x2_float)
[…]
