+++ pyccel/ast/core.py
@@ -467,7 +467,7 @@ class Assign(Basic):
         return False
 
 #------------------------------------------------------------------------------
-class Allocate(Basic):
+class  Allocate(Basic):
     """
     Represents memory allocation (usually of an array) for code generation.
     This is relevant to low-level target languages, such as C or Fortran,
