### 2023-01-02 6cb258d
changes comment in C code ndarrays tests
