+++ tests/ndarrays/test_ndarrays.c
@@ -165,7 +165,7 @@ int32_t test_indexing_int64(void)
 
     x = array_create(2, m_1_shape, nd_int64, false, order_c);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 26;
     my_assert(index , c_index, "testing the strides");
@@ -194,7 +194,7 @@ int32_t test_indexing_int64_order_f(void)
 
     x = array_create(2, m_1_shape, nd_int64, false, order_f);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 13;
     my_assert(index , c_index, "testing the strides");
@@ -223,7 +223,7 @@ int32_t test_indexing_int32(void)
 
     x = array_create(2, m_1_shape, nd_int32, false, order_c);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 26;
     my_assert(index , c_index, "testing the strides");
@@ -252,7 +252,7 @@ int32_t test_indexing_int32_order_f(void)
 
     x = array_create(2, m_1_shape, nd_int32, false, order_f);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 13;
     my_assert(index , c_index, "testing the strides");
@@ -282,7 +282,7 @@ int32_t test_indexing_int16(void)
 
     x = array_create(2, m_1_shape, nd_int16, false, order_c);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 26;
     my_assert(index , c_index, "testing the strides");
@@ -311,7 +311,7 @@ int32_t test_indexing_int16_order_f(void)
 
     x = array_create(2, m_1_shape, nd_int16, false, order_f);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 13;
     my_assert(index , c_index, "testing the strides");
@@ -340,7 +340,7 @@ int32_t test_indexing_int8(void)
 
     x = array_create(2, m_1_shape, nd_int8, false, order_c);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 26;
     my_assert(index , c_index, "testing the strides");
@@ -369,7 +369,7 @@ int32_t test_indexing_int8_order_f(void)
 
     x = array_create(2, m_1_shape, nd_int8, false, order_f);
     memcpy(x.raw_data, m_1, x.buffer_size);
-    // testing the index [3, 2]
+    // testing the strides for index [3, 2]
     index = 3 * x.strides[0] + 2 * x.strides[1];
     c_index = 13;
     my_assert(index , c_index, "testing the strides");
@@ -398,7 +398,7 @@ int32_t test_indexing_double(void)
 
[…]
