### 2023-01-02 b71dd36
Clarify 'these functions'
