### 2023-01-02 f04b412
comment semantic device/kernel conditionals
