+++ setup.cfg
@@ -13,7 +13,7 @@ long_description_content_type = text/markdown
 [options]
 packages = find:
 install_requires =
-    numpy
+    numpy >= 1.16, < 1.24
     sympy>=1.2
     termcolor
     textx>=2.2
