+++ pyccel/parser/semantic.py
@@ -2631,11 +2631,10 @@ class SemanticParser(BasicParser):
 
             new_expr = Assign(l, r)
 
+            if isinstance(expr, AugAssign):
+                new_expr = AugAssign(l, expr.op, r)
             if is_pointer_i:
-                if isinstance(expr, AugAssign):
-                    new_expr = AugAssign(l, expr.op, r)
-                else:
-                    new_expr = AliasAssign(l, r)
+                new_expr = AliasAssign(l, r)
 
             elif isinstance(expr, AugAssign):
                 new_expr = AugAssign(l, expr.op, r)
