### 2023-01-02 7f2400c
checking the return value
