### 2023-01-02 5ff5c22
Use managed_memory to allocate shape and strides
