### 2023-01-03 ca11636
started GPU_ddocs
