### 2023-01-03 d451336
update test file of deallocate
