+++ tests/internal/scripts/ccuda/free_pointer.py
@@ -1,3 +1,5 @@
+# pylint: disable=missing-function-docstring, disable=unused-variable, missing-module-docstring
+
 from pyccel.decorators import kernel
 from pyccel import cuda
 import math
@@ -11,9 +13,7 @@ def func(arr:'int[:]'):
     arr[i] = math.pow(arr[i], 2)
 
 if __name__ == '__main__':
-    a = cuda.array([1,2,3,4], memory_location='device')
-    b = cuda.array([3,4,5,6], memory_location='host')
-    c = b
-    print(b)
+    a = cuda.array([1,2,3,4], memory_location='device')   
     func[1,4](a)
+    c = a
     cuda.synchronize()
