+++ developer_docs/GPU_getting_started.md
@@ -1,8 +1,8 @@
 # Getting started GPU
 
-## Decorators:
+## Decorators
 
-* Kernel
+  * Kernel
 
     In Pyccel, the `@kernel` decorator is used to indicate that a function should be treated as a CUDA kernel. A CUDA kernel is a function that is executed on a GPU and is typically used to perform a parallel computation over a large dataset.
 
@@ -27,7 +27,7 @@
     tn = 64   # size of the block
     my_kernel[ng, tn](x, y, out)
     ```
-* Device
+  * Device
 
     The `@device` decorator is similar to the `@kernel` decorator, but indicates that the function should be compiled and executed on the GPU as a device function, rather than a kernel.
     
@@ -50,16 +50,15 @@
         y[i] = my_device_function(x[i])
     ```
 
-## Built-in variables:
+## Built-in variables
 
-* cuda.threadIdx(dim): Returns the index of the current thread within the block in a given dimension dim. dim is an integer between 0 and the number of dimensions of the thread block minus one.
+  * cuda.threadIdx(dim): Returns the index of the current thread within the block in a given dimension dim. dim is an integer between 0 and the number of dimensions of the thread block minus one.
 
-* cuda.blockIdx(dim): Returns the index of the current block within the grid in a given dimension dim.
-dim is an integer between 0 and the number of dimensions of the grid minus one.
+  * cuda.blockIdx(dim): Returns the index of the current block within the grid in a given dimension dim.dim is an integer between 0 and the number of dimensions of the grid minus one.
 
-* cuda.blockDim(dim): Returns the size of the thread block in a given dimension dim. dim is an integer between 0 and the number of dimensions of the thread block minus one.
+  * cuda.blockDim(dim): Returns the size of the thread block in a given dimension dim. dim is an integer between 0 and the number of dimensions of the thread block minus one.
 
-* cuda.gridDim(dim): Returns the size of the grid in a given dimension dim. dim is an integer between 0 and the number of dimensions of the grid minus one.
+  * cuda.gridDim(dim): Returns the size of the grid in a given dimension dim. dim is an integer between 0 and the number of dimensions of the grid minus one.
 
 > These built-in variables are provided by the CUDA runtime and are available to all CUDA kernels, regardless of the programming language being used. They are used to obtain information about the execution configuration of the kernel, such as the indices and dimensions of the threads and blocks. They are an important part of the CUDA programming model and are often used to compute global indices and other information about the execution configuration of the kernel.
 
