+++ tests/internal/scripts/ccuda/free_pointer.py
@@ -1,8 +1,8 @@
 # pylint: disable=missing-function-docstring, disable=unused-variable, missing-module-docstring
 
+import math
 from pyccel.decorators import kernel
 from pyccel import cuda
-import math
 
 @kernel
 def func(arr:'int[:]'):
@@ -13,7 +13,7 @@ def func(arr:'int[:]'):
     arr[i] = math.pow(arr[i], 2)
 
 if __name__ == '__main__':
-    a = cuda.array([1,2,3,4], memory_location='device')   
+    a = cuda.array([1,2,3,4], memory_location='device')
     func[1,4](a)
     c = a
     cuda.synchronize()
