+++ ci_tools/coverage_analysis_tools.py
@@ -104,8 +104,8 @@ def allow_untested_error_calls(untested):
     for f,line_nums in untested.items():
         with open(f, encoding="utf-8") as filename:
             f_lines = filename.readlines()
-        untested_lines = [f_lines[i-1].strip() for i in line_nums]
-        lines = [l for l in untested_lines if not (l.startswith('raise ') or l.startswith('errors.report(') or l.startswith('return errors.report('))]
+        untested_lines = [(i, f_lines[i-1].strip()) for i in line_nums]
+        lines = [i for i,l in untested_lines if not (l.startswith('raise ') or l.startswith('errors.report(') or l.startswith('return errors.report('))]
         if len(lines):
             reduced_untested[f] = lines
 
