+++ tests/cuda_test/test_kernel_semantic.py
@@ -173,3 +173,51 @@ def test_missing_kernel_config(language):
     error_info = [*errors.error_info_map.values()][0][0]
     assert error_info.symbol.func_name == 'kernel_call'
     assert MISSING_KERNEL_CONFIGURATION == error_info.message
+
+@pytest.mark.parametrize( 'language', [
+        pytest.param("ccuda", marks = pytest.mark.ccuda)
+    ]
+)
+def test_invalid_block_number(language):
+    def invalid_block_number():
+        @kernel
+        def kernel_call():
+            pass
+        blocks_per_grid = 5.0
+        kernel_call[blocks_per_grid, 1]()
+
+    errors = Errors()
+
+    with pytest.raises(PyccelSemanticError):
+        epyccel(invalid_block_number, language=language)
+
+    assert errors.has_errors()
+    assert errors.num_messages() == 1
+
+    error_info = [*errors.error_info_map.values()][0][0]
+    assert error_info.symbol.func == 'kernel_call'
+    assert  INVALID_KERNEL_CALL_BP_GRID == error_info.message
+
+@pytest.mark.parametrize( 'language', [
+        pytest.param("ccuda", marks = pytest.mark.ccuda)
+    ]
+)
+def test_invalid_thread_per_block(language):
+    def invalid_thread_per_block():
+        @kernel
+        def kernel_call():
+            pass
+        threads_per_block = 5.0
+        kernel_call[1, threads_per_block]()
+
+    errors = Errors()
+
+    with pytest.raises(PyccelSemanticError):
+        epyccel(invalid_thread_per_block, language=language)
+
+    assert errors.has_errors()
+    assert errors.num_messages() == 1
+
+    error_info = [*errors.error_info_map.values()][0][0]
+    assert error_info.symbol.func == 'kernel_call'
+    assert  INVALID_KERNEL_CALL_TP_BLOCK == error_info.message
