### 2023-01-05 ea48de9
Fix CoverageChecker bugs
