### 2023-01-06 d472a7c
Do a copy for host arrays
