### 2023-01-06 2440d2a
Add coverage to cuda tests
