+++ .github/workflows/Github_pytest.yml
@@ -2,7 +2,7 @@ name: Pyccel tests
 
 on:
   pull_request:
-    branches: [ cuda_main_temp ]
+    branches: [ master, cuda_main_temp ]
 
 jobs:
   Linux:
@@ -32,17 +32,11 @@ jobs:
       - name: Collect coverage information
         continue-on-error: True
         uses: ./.github/actions/coverage_collection
-      - name: Run codacy-coverage-reporter
-        uses: codacy/codacy-coverage-reporter-action@master
-        continue-on-error: True
-        with:
-          project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
-          coverage-reports: cobertura.xml
       - name: Save code coverage report
         uses: actions/upload-artifact@v3
         with:
           name: coverage-artifact
-          path: cobertura.xml
+          path: .coverage
           retention-days: 1
 
   Windows:
@@ -95,6 +89,33 @@ jobs:
       - name: Parallel tests with pytest
         uses: ./.github/actions/pytest_parallel
 
+  Cuda:
+
+    runs-on: ubuntu-latest
+    container: nvidia/cuda:11.7.1-devel-ubuntu20.04
+    if: github.event.pull_request.base.ref == 'cuda_main_temp'
+    steps:
+      - uses: actions/checkout@v2
+      - name: CUDA Version
+        run: nvcc --version # cuda install check
+      - name: Install dependencies
+        uses: ./.github/actions/linux_install
+      - name: Install python dependencies
+        uses: ./.github/actions/pip_installation
+      - name: Coverage install
+        uses: ./.github/actions/coverage_install
+      - name: Ccuda tests with pytest
+        uses: ./.github/actions/pytest_run_cuda
+      - name: Collect coverage information
+        continue-on-error: True
+        uses: ./.github/actions/coverage_collection
+      - name: Save code coverage report
+        uses: actions/upload-artifact@v3
+        with:
+          name: cuda-coverage-artifact
+          path: .coverage
+          retention-days: 1
+
   Linter:
 
     runs-on: ubuntu-latest
@@ -115,10 +136,43 @@ jobs:
             python -m pylint --rcfile=.pylintrc pyccel/parser/semantic.py
         shell: bash
 
+  CoverageCollection:
+
+    runs-on: ubuntu-latest
+    needs: [Linux, Cuda]
+
+    steps:
+      - name: Collect coverage information
+        uses: actions/download-artifact@v3
+        with:
+          name: coverage-artifact
+      - name: Rename coverage file
+        run: mv .coverage .coverage-linux
+      - name: Collect coverage information
+        uses: actions/download-artifact@v3
+        if: 
+        with:
+          name: cuda-coverage-artifact
+      - name: Rename coverage file
+        if: ${{ success() }}
+        run: mv .coverage .coverage-cuda
+      - name: Generate coverage report
+        run: |
+            ls -a
+            coverage combine
+            ls -a
+            coverage xml
+      - name: Run codacy-coverage-reporter
+        uses: codacy/codacy-coverage-reporter-action@master
+        continue-on-error: True
+        with:
+          project-token: ${{ secrets.CODACY_PROJECT_TOKEN }}
+          coverage-reports: cobertura.xml
+
   CoverageChecker:
 
     runs-on: ubuntu-latest
-    needs: Linux
+    needs: [CoverageCollection]
 
     steps:
       - uses: actions/checkout@v3
