+++ .github/workflows/Github_pytest.yml
@@ -140,6 +140,7 @@ jobs:
 
     runs-on: ubuntu-latest
     needs: [Linux, Cuda]
+    if: ${{ always() }}
 
     steps:
       - name: Collect coverage information
