+++ .github/workflows/Github_pytest.yml
@@ -143,6 +143,13 @@ jobs:
     if: ${{ always() }}
 
     steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.7
+      - name: Coverage install
+        run: python -m pip install coverage
       - name: Collect coverage information
         uses: actions/download-artifact@v3
         with:
