+++ .github/workflows/Github_pytest.yml
@@ -155,7 +155,7 @@ jobs:
         with:
           name: coverage-artifact
       - name: Rename coverage file
-        run: mv .coverage .coverage-linux
+        run: mv .coverage .coverage.linux
       - name: Collect coverage information
         uses: actions/download-artifact@v3
         if: 
@@ -163,7 +163,7 @@ jobs:
           name: cuda-coverage-artifact
       - name: Rename coverage file
         if: ${{ success() }}
-        run: mv .coverage .coverage-cuda
+        run: mv .coverage .coverage.cuda
       - name: Generate coverage report
         if: ${{ always() }}
         run: |
