+++ .github/workflows/Github_pytest.yml
@@ -148,6 +148,10 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: 3.7
+      - name: Install dependencies
+        uses: ./.github/actions/linux_install
+      - name: Install python dependencies
+        uses: ./.github/actions/pip_installation
       - name: Coverage install
         run: python -m pip install coverage
       - name: Collect coverage information
@@ -167,9 +171,7 @@ jobs:
       - name: Generate coverage report
         if: ${{ always() }}
         run: |
-            ls -a
             coverage combine
-            ls -a
             coverage xml
       - name: Run codacy-coverage-reporter
         uses: codacy/codacy-coverage-reporter-action@master
