### 2023-01-06 3485879
Trigger for cuda_main and cuda_main_temp
