+++ .github/workflows/Github_pytest.yml
@@ -143,7 +143,7 @@ jobs:
 
     runs-on: ubuntu-latest
     needs: [Linux, Cuda]
-    if: ${{ always() }}
+    if: ${{ always() && needs.Linux.result != 'failed' }}
 
     steps:
       - uses: actions/checkout@v3
