### 2023-01-06 8f51b28
Try alternative definition
