+++ .github/actions/coverage_install/action.yml
@@ -11,6 +11,10 @@ runs:
       run: |
         INSTALL_DIR=$(cd tests; python -c "import pyccel; print(pyccel.__path__[0])")
         SITE_DIR=$(python -c 'import sysconfig; print(sysconfig.get_paths()["purelib"])')
+        echo "INSTALL_DIR = ${INSTALL_DIR}"
+        echo "SITE_DIR = ${SITE_DIR}"
+        SITE_DIR=$(dirname ${INSTALL_DIR})
+        echo "SITE_DIR = ${SITE_DIR}"
         echo -e "import coverage; coverage.process_startup()" > ${SITE_DIR}/pyccel_cov.pth
         echo -e "[run]\nparallel = True\nsource = ${INSTALL_DIR}\ndata_file = $(pwd)/.coverage\n[report]\ninclude = ${INSTALL_DIR}/*\n[xml]\noutput = cobertura.xml" > .coveragerc
         echo "SITE_DIR=${SITE_DIR}" >> $GITHUB_ENV
