+++ .github/actions/coverage_install/action.yml
@@ -10,9 +10,6 @@ runs:
     - name: Directory Creation
       run: |
         INSTALL_DIR=$(cd tests; python -c "import pyccel; print(pyccel.__path__[0])")
-        SITE_DIR=$(python -c 'import sysconfig; print(sysconfig.get_paths()["purelib"])')
-        echo "INSTALL_DIR = ${INSTALL_DIR}"
-        echo "SITE_DIR = ${SITE_DIR}"
         SITE_DIR=$(dirname ${INSTALL_DIR})
         echo "SITE_DIR = ${SITE_DIR}"
         echo -e "import coverage; coverage.process_startup()" > ${SITE_DIR}/pyccel_cov.pth
