### 2023-01-06 61c7a4d
Activate cuda tests
