+++ .github/workflows/Cuda_pytest.yml
@@ -11,6 +11,13 @@ jobs:
     container: nvidia/cuda:11.7.1-devel-ubuntu20.04
     steps:
       - uses: actions/checkout@v2
+      - name: Prepare docker
+        run: |
+          apt update && apt install sudo
+          TZ=Europe/France
+          ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
+          DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends tzdata
+        shell: bash
       - name: CUDA Version
         run: nvcc --version # cuda install check
       - name: Install dependencies
