### 2023-01-06 68bb801
Update pyccel/parser/semantic.py
Co-authored-by: EmilyBourne <louise.bourne@gmail.com>
