+++ .github/workflows/Github_pytest.yml
@@ -147,7 +147,7 @@ jobs:
 
     runs-on: ubuntu-latest
     needs: [Linux, Cuda]
-    if: ${{ always() && needs.Linux.result == 'success' }}
+    if: ${{ always() && needs.Linux.result == 'success' && needs.Cuda.result != 'failure' }}
 
     steps:
       - uses: actions/checkout@v3
@@ -196,7 +196,7 @@ jobs:
 
     runs-on: ubuntu-latest
     needs: [CoverageCollection]
-    if: ${{ always() && needs.CoverageCollection.result == 'success' }}
+    if: ${{ always() && needs.CoverageCollection.result == 'success' && needs.Cuda.result != 'failure' }}
 
     steps:
       - uses: actions/checkout@v3
