### 2023-01-06 e5d7ba0
Update variable search method
