### 2023-01-06 d2c3498
raise error for calling non-kernel functions with parameters
