+++ pyccel/errors/messages.py
@@ -165,4 +165,5 @@ NON_KERNEL_FUNCTION_CUDA_VAR = 'Cuda internal variables should only be used in K
 INVALID_KERNEL_CALL_BP_GRID = 'Invalid Block per grid parameter for Kernel call'
 INVALID_KERNEL_CALL_TP_BLOCK = 'Invalid Thread per Block parameter for Kernel call'
 MISSING_KERNEL_CONFIGURATION = 'Kernel launch configuration not specified'
+INVALID_FUNCTION_CALL = 'Invalid call for a non-kernel function'
 
