+++ pyccel/parser/semantic.py
@@ -404,7 +404,7 @@ class SemanticParser(BasicParser):
         self._allocs.pop()
         return deallocs
 
-    def _infere_type(self, expr, **settings):
+    def _infer_type(self, expr, **settings):
         """
         type inference for expressions
         """
@@ -465,7 +465,7 @@ class SemanticParser(BasicParser):
             return d_var
 
         elif isinstance(expr, Duplicate):
-            d = self._infere_type(expr.val, **settings)
+            d = self._infer_type(expr.val, **settings)
 
             # TODO must check that it is consistent with pyccel's rules
             # TODO improve
@@ -516,7 +516,7 @@ class SemanticParser(BasicParser):
             return d_var
 
         elif isinstance(expr, IfTernaryOperator):
-            return self._infere_type(expr.args[0][1].body[0])
+            return self._infer_type(expr.args[0][1].body[0])
 
         elif isinstance(expr, PythonRange):
 
@@ -552,7 +552,7 @@ class SemanticParser(BasicParser):
             return d_var
 
         elif isinstance(expr, GeneratorComprehension):
-            return self._infere_type(expr.lhs)
+            return self._infer_type(expr.lhs)
 
         else:
             type_name = type(expr).__name__
@@ -805,10 +805,10 @@ class SemanticParser(BasicParser):
             args, kwargs = split_positional_keyword_arguments(*args)
             for a in args:
                 if getattr(a,'dtype',None) == 'tuple':
-                    self._infere_type(a, **settings)
+                    self._infer_type(a, **settings)
             for a in kwargs.values():
                 if getattr(a,'dtype',None) == 'tuple':
-                    self._infere_type(a, **settings)
+                    self._infer_type(a, **settings)
 
             try:
                 new_expr = func(*args, **kwargs)
@@ -881,7 +881,7 @@ class SemanticParser(BasicParser):
 
         rhs : Variable
             The value assigned to the lhs. This is required to call
-            self._infere_type recursively for tuples
+            self._infer_type recursively for tuples
 
         arr_in_multirets : bool
             If True, the variable that will be created is an array
@@ -906,7 +906,7 @@ class SemanticParser(BasicParser):
             elem_d_lhs_ref = None
             for i,r in enumerate(iterable):
                 elem_name = self.scope.get_new_name( name + '_' + str(i) )
-                elem_d_lhs = self._infere_type( r )
+                elem_d_lhs = self._infer_type( r )
 
                 if not arr_in_multirets:
                     self._ensure_target( r, elem_d_lhs )
@@ -1108,7 +1108,7 @@ class SemanticParser(BasicParser):
             # Variable already exists
             else:
 
-                self._ensure_infered_type_matches_existing(dtype, d_var, var, is_augassign, new_expressions, rhs)
+                self._ensure_inferred_type_matches_existing(dtype, d_var, var, is_augassign, new_expressions, rhs)
 
                 # in the case of elemental, lhs is not of the same dtype as
                 # var.
@@ -1157,7 +1157,7 @@ class SemanticParser(BasicParser):
 
         return lhs
 
-    def _ensure_infered_type_matches_existing(self, dtype, d_var, var, is_augassign, new_expressions, rhs):
+    def _ensure_inferred_type_matches_existing(self, dtype, d_var, var, is_augassign, new_expressions, rhs):
         """
         Ensure that the inferred type of the new variable, matches the existing variable (which has the
         same name). If this is not the case then errors are raised preventing pyccel reaching the codegen
@@ -1170,7 +1170,7 @@ class SemanticParser(BasicParser):
         dtype : DataType
                 The inferred DataType
         d_var : dict
-                The inferred information about the variable. Usually created by the _infere_type function
[…]
