### 2023-01-07 9ea3abf
Revert change on action.yml
