### 2023-01-07 6f16b4e
Use get to check if array's memory location
