### 2023-01-07 b8643d2
Use tuple instead of list for variable
