### 2023-01-07 be2fd09
Remove _attribute_nodes from CupyRavel
