### 2023-01-07 3fb754f
Use tuple instead of list for variable in test
