### 2023-01-07 4653b6b
add test for invalid function call error
