+++ tests/cuda_test/test_kernel_semantic.py
@@ -6,7 +6,7 @@ import numpy as np
 from pyccel.epyccel import epyccel
 from pyccel.decorators import stack_array, types, kernel
 from pyccel.errors.errors import Errors, PyccelSemanticError
-from pyccel.errors.messages import (KERNEL_STACK_ARRAY_ARG, MISSING_KERNEL_CONFIGURATION,
+from pyccel.errors.messages import (INVALID_FUNCTION_CALL, KERNEL_STACK_ARRAY_ARG, MISSING_KERNEL_CONFIGURATION,
                                     NON_KERNEL_FUNCTION_CUDA_VAR,
                                     INVALID_KERNEL_CALL_BP_GRID,
                                     INVALID_KERNEL_CALL_TP_BLOCK
@@ -221,3 +221,26 @@ def test_invalid_thread_per_block(language):
     error_info = [*errors.error_info_map.values()][0][0]
     assert error_info.symbol.func == 'kernel_call'
     assert  INVALID_KERNEL_CALL_TP_BLOCK == error_info.message
+
+
+@pytest.mark.parametrize( 'language', [
+        pytest.param("ccuda", marks = pytest.mark.ccuda)
+    ]
+)
+def test_invalid_function_call(language):
+    def invalid_function_call():
+        def function_call():
+            pass
+        function_call[1, 2]()
+
+    errors = Errors()
+
+    with pytest.raises(PyccelSemanticError):
+        epyccel(invalid_function_call, language=language)
+
+    assert errors.has_errors()
+    assert errors.num_messages() == 1
+
+    error_info = [*errors.error_info_map.values()][0][0]
+    assert error_info.symbol.func == 'function_call'
+    assert  INVALID_FUNCTION_CALL == error_info.message
