### 2023-01-07 #4991 @boegel
> 👋 @boegel - how is your review coming?

@danielskatz My apologies for the delay!

I've worked through most of it, but I'm still playing with the pyccel benchmarks to see if I can reproduce the results included in the paper.
I have some suggestion for improvements in that specific section too: the paper doesn't mention what type of hardware the benchmark results were collected on, or how many cores were used.
In additional the use of `-ffast-math` is a bit concerning (see https://simonbyrne.github.io/notes/fastmath/).

I should open an issue on this in the pyccel repo, right?
