### 2023-01-07 bcfc506
Remove useless code
