### 2023-01-07 aafa425
Remove dead code
