### 2023-01-07 def5e4e
Remove unnecessary arg property getter CupyRavel
