### 2023-01-07 f9b3dfe
issue_1254 - fix the failing tests
