+++ tests/epyccel/test_epyccel_complex_func.py
@@ -46,7 +46,7 @@ def test_create_complex_var__complex_float(language):
 
     a = complex(randint(100), randint(100))
     b = rand()*100
-    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
 
 def test_create_complex_var__complex_complex(language):
     f = mod.create_complex_var__complex_complex
@@ -54,7 +54,7 @@ def test_create_complex_var__complex_complex(language):
 
     a = complex(randint(100), randint(100))
     b = complex(randint(100), randint(100))
-    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
 
 def test_create_complex__int_int(language):
     f = mod.create_complex__int_int
@@ -75,35 +75,35 @@ def test_create_complex__float_float(language):
     f_epyc = epyccel(f, language = language)
 
     a = rand()*100
-    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_create_complex_0__float_float(language):
     f = mod.create_complex_0__float_float
     f_epyc = epyccel(f, language = language)
 
     a = rand()*100
-    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_create_complex__complex_complex(language):
     f = mod.create_complex__complex_complex
     f_epyc = epyccel(f, language = language)
 
     a = complex(randint(100), randint(100))
-    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_cast_complex_1(language):
     f = mod.cast_complex_1
     f_epyc = epyccel(f, language = language)
 
     a = np.complex64(complex(randint(100), randint(100)))
-    assert np.isclose(f_epyc(a), f(a), rtol = 1e-7, atol = 1e-8)
+    assert np.allclose(f_epyc(a), f(a), rtol = 1e-7, atol = 1e-8)
 
 def test_cast_complex_2(language):
     f = mod.cast_complex_2
     f_epyc = epyccel(f, language = language)
 
     a = np.complex128(complex(randint(100), randint(100)))
-    assert np.isclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a), f(a), rtol=RTOL, atol=ATOL)
 
 def test_cast_float_complex(language):
     f = mod.cast_float_complex
@@ -111,4 +111,4 @@ def test_cast_float_complex(language):
 
     a = rand()*100
     b = complex(randint(100), randint(100))
-    assert np.isclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
+    assert np.allclose(f_epyc(a,b), f(a,b), rtol=RTOL, atol=ATOL)
