### 2023-01-07 pyccel/pyccel#1295 @boegel
[JOSS] Missing details for benchmark results
(cfr. https://github.com/openjournals/joss-reviews/issues/4991)

The submission to JOSS (cfr. [joss-paper](https://github.com/pyccel/pyccel/tree/joss-paper/joss-paper), submitted by @yguclu ) includes some nice benchmark results, collected via [pyccel-benchmarks](https://github.com/pyccel/pyccel-benchmarks).

There are some important details missing about the experimental setup though, in particular:

* operating system + version
* hardware details (mainly which exact CPU)
* how many cores are in the system, and how many were used for running the benchmarks

Can this information be added to the paper?
It would be very helpful to interpret the results.

In addition, the use of `-ffast-math` when benchmarking Pythran and Pyccel is a bit concerning, see for example https://simonbyrne.github.io/notes/fastmath/ .
Is `-ffast-math` critical to get good results? If not, would it make sense to re-run the benchmarks without it, and include those results in the JOSS paper instead?
