### 2023-01-07 6daad4f
set view strides and shape starting dimension to j-1 instead of j
