### 2023-01-07 f33bdc1
Remove dead code
