+++ developer_docs/semantic_stage.md
@@ -147,6 +147,10 @@ The attribute `SemanticParser._additional_exprs` exists to hold these expression
 This object is a list of lists which is initialised and inserted in `_visit_CodeBlock`.
 A list of lists is necessary in case a CodeBlock can be found inside another (e.g. `a = [i for i in range]` where a `CodeBlock` contains the `Assign`, but another exists inside the `For` loop).
 
+In order to avoid problems arising from forgetfulness we try to add additional objects in the most general place possible.
+For example, allocation occurs in the function `SemanticParser._assign_lhs_variable`.
+Declarations are created in the printer when needed from the scope variables (this allows each language to place the decorators in the most appropriate location).
+
 ## Object Tree
 
 All the objects in the Pyccel AST inherit from [`pyccel.ast.basic.Basic`](../pyccel/ast/basic.py).
