### 2023-01-08 abc1a07
Update docs. enumerate/zip/map can't be passed to list or tuple
