+++ .github/workflows/spelling.yml
@@ -1,0 +1,14 @@
+name: Spellcheck Action
+on:
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+    name: Spellcheck
+    runs-on: ubuntu-latest
+    steps:
+    # The checkout step
+    - uses: actions/checkout@v3
+    - uses: rojopolis/spellcheck-github-actions@0.29.0
+      name: Spellcheck
