+++ .github/workflows/spelling.yml
@@ -4,7 +4,7 @@ on:
     branches: [ master ]
 
 jobs:
-  build:
+  Spelling:
     name: Spellcheck
     runs-on: ubuntu-latest
     steps:
