+++ .dict_custom.txt
@@ -1,0 +1,4 @@
+pyccel
+pythran
+numba
+numpy
