+++ .pyspelling.yml
@@ -1,0 +1,18 @@
+matrix:
+- name: markdown
+  sources:
+  - README.md
+  - tutorial/*.md
+  - developer_docs/*.md
+  aspell:
+    lang: en
+    d: en_GB
+  dictionary:
+    wordlists:
+    - .dict_custom.txt
+    output: build/dictionary/python.dic
+  pipeline:
+  - pyspelling.filters.markdown:
+  - pyspelling.filters.html:
+      ignores:
+      - code
