+++ .github/workflows/spelling.yml
@@ -5,10 +5,25 @@ on:
 
 jobs:
   Spelling:
-    name: Spellcheck
     runs-on: ubuntu-latest
     steps:
-    # The checkout step
-    - uses: actions/checkout@v3
-    - uses: rojopolis/spellcheck-github-actions@0.29.0
-      name: Spellcheck
+    - uses: actions/checkout@v1
+    - name: Set up Python 3.7
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.7
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip setuptools
+        python -m pip install pyspelling
+        # Install any additional libraries required: additional plugins, documentation building libraries, etc.
+    - name: Install Aspell
+      run: |
+        sudo apt-get install aspell aspell-en
+    - name: Build documents
+      run: |
+        # Perform any documentation building that might be required
+    - name: Spell check
+       run: |
+         python -m pyspelling > $GITHUB_STEP_SUMMARY
+
