+++ .dict_custom.txt
@@ -1,4 +1,42 @@
-pyccel
-pythran
+Pyccel
+Pythran
 numba
-numpy
+NumPy
+BLAS
+LAPACK
+MPI
+OpenMP
+Fortran
+pyccelise
+allocatable
+deallocate
+conda
+PyPI
+CentOS
+RHEL
+executables
+linux
+macos
+DLL
+DLLs
+MPICH
+openSUSE
+Xcode
+SIMD
+runtime
+pragma
+API
+APIs
+Pyccel's
+PGI
+GFortran
+DNF
+Homebrew
+CLT
+SDK
+github
+OpenMPI
+JSON
+CPython
+IPython
+HPC
