+++ .pyspelling.yml
@@ -10,9 +10,17 @@ matrix:
   dictionary:
     wordlists:
     - .dict_custom.txt
+    encoding: utf-8
     output: build/dictionary/python.dic
   pipeline:
   - pyspelling.filters.markdown:
+      markdown_extensions:
+      - pymdownx.superfences:
+      - pymdownx.magiclink:
   - pyspelling.filters.html:
+      comments: false
       ignores:
       - code
+      - pre
+      - a
+  default_encoding: utf-8
