+++ tutorial/compiler.md
@@ -2,16 +2,16 @@
 ## Compilers supported by Pyccel
 
 Pyccel provides default compiler settings for 4 different compiler families:
--   **GNU** : gcc / gfortran
--   **intel** : icc / ifort
--   **PGI** : pgcc / pgfortran
--   **nvidia** : nvc / nvfort
+-   **GNU** : `gcc` / `gfortran`
+-   **intel** : `icc` / `ifort`
+-   **PGI** : `pgcc` / `pgfortran`
+-   **nvidia** : `nvc` / `nvfort`
 
 **Warning** : The **GNU** compiler is currently the only compiler which is tested regularly
 
 ## Specifying a compiler
 
-The default compiler family is **GNU**. To use a different compiler, the compiler family should be passed to either pyccel or epyccel.
+The default compiler family is **GNU**. To use a different compiler, the compiler family should be passed to either `pyccel` or `epyccel`.
 E.g.
 ```shell
 pyccel example.py --compiler=intel
@@ -23,7 +23,7 @@ epyccel(my_func, compiler='intel')
 
 ## User-defined compiler
 
-The user can also define their own compiler in a json file. To use this definition, the location of the json file must be passed to the _compiler_ argument. The json file must define the following:
+The user can also define their own compiler in a JSON file. To use this definition, the location of the JSON file must be passed to the _compiler_ argument. The JSON file must define the following:
 
 -   **exec** : The name of the executable
 -   **mpi\_exec** : The name of the mpi executable
@@ -37,7 +37,7 @@ The user can also define their own compiler in a json file. To use this definiti
 -   **libdirs** : A list of library directories necessary for compiling \[optional\]
 -   **includes** : A list of include directories necessary for compiling \[optional\]
   
-In addition, for each accelerator (mpi/openmp/openacc/python) that you will use the json file must define the following:
+In addition, for each accelerator (mpi/openmp/openacc/python) that you will use the JSON file must define the following:
   
 -   **flags** : A list of flags used to impose the expected language standard \[optional\]
 -   **libs** : A list of libraries necessary for compiling \[optional\]
@@ -46,7 +46,7 @@ In addition, for each accelerator (mpi/openmp/openacc/python) that you will use 
 
 Python is considered to be an accelerator and must additionally specify shared\_suffix.
 
-The default compilers can provide examples compatible with your system once pyccel has been executed at least. To export the json file describing your setup, use the `--export-compile-info` flag and provide a target file name.
+The default compilers can provide examples compatible with your system once pyccel has been executed at least. To export the JSON file describing your setup, use the `--export-compile-info` flag and provide a target file name.
 E.g.
 ```shell
 pyccel --compiler=PGI --language=c --export-compile-info=icc.json
