+++ tutorial/ndarrays.md
@@ -315,14 +315,14 @@ Some examples:
 
 -   Supported [types](https://numpy.org/devdocs/user/basics.types.html):
 
-    bool, int, int8, int16, int32, int64, float, float32, float64, complex64 and complex128. They can be used as cast functions too.
+    `bool`, `int`, `int8`, `int16`, `int32`, `int64`, `float`, `float32`, `float64`, `complex`, `complex64` and `complex128`. They can be used as cast functions too.
 
-    Note: np.bool, np.int and np.float are just aliases to the Python native types, and are considered as a deprecated way to work with Python built-in types in NumPy.
+    Note: `np.bool`, `np.int`, `np.float` and `np.complex` are just aliases to the Python native types, and are considered as a deprecated way to work with Python built-in types in NumPy.
 
 -   Properties:
 
-    -   real, imag, shape, amax, amin
+    -   `real`, `imag`, `shape`, `amax`, `amin`
 
 -   Methods:
 
-    -   sum
+    -   `sum`
