+++ .github/workflows/spelling.yml
@@ -23,7 +23,7 @@ jobs:
         sudo apt-get install aspell aspell-en
     - name: Spell check
       run: |
-         python -m pyspelling > misspellings.txt
+         python -m pyspelling > misspellings.txt || true
          cat misspellings.txt
          python ci_tools/summarise_pyspelling.py misspellings.txt $GITHUB_STEP_SUMMARY
 
