+++ tutorial/header-files.md
@@ -82,7 +82,7 @@ def print_fib(x):
 To compile this file we execute the following command `pyccel test_funcs.py --libs=funcs --libdir=$PWD`, this will create the shared library `test_funcs.so`
 
 ## Pickling header files
-Parsing a large Pyccel header file with hundreds of function declarations may require a significant amount of time, therefore it is important that this process is only done once when pyccelizing multiple Python source files in a large project.
+Parsing a large Pyccel header file with hundreds of function declarations may require a significant amount of time, therefore it is important that this process is only done once when pyccelising multiple Python source files in a large project.
 
 To this end, Pyccel uses the [pickle](https://docs.python.org/3/library/pickle.html) Python module to store the result of the parser to a `.pyccel` binary file, which is created in the same directory as the header file.
 Afterwards Pyccel will load the precompiled parser from the `.pyccel` file, instead of parsing the header file again.
