+++ tutorial/numpy-functions.md
@@ -545,19 +545,19 @@ In Pyccel we try to support the Numpy functions which developers use the most.. 
 
 -   Supported [math functions](https://numpy.org/doc/stable/reference/routines.math.html) (optional parameters are not supported):
 
-    sqrt, abs, sin, cos, exp, log, tan, arcsin, arccos, arctan, arctan2, sinh, cosh, tanh, arcsinh, arccosh and
-    arctanh.
+    `sqrt`, `abs`, `sin`, `cos`, `exp`, `log`, `tan`, `arcsin`, `arccos`, `arctan`, `arctan2`, `sinh`, `cosh`, `tanh`, `arcsinh`, `arccosh` and
+    `arctanh`.
 
 -   Supported [array creation routines](https://numpy.org/doc/stable/reference/routines.array-creation.html) (fully supported):
 
-    -   empty, full, ones, zeros, arange (`like` parameter is not supported).
-    -   empty_like, full_like, and zeros_like, ones_like (`subok` parameter is not supported).
-    -   rand, randint
-    -   where, count_nonzero (fortran only)
-    -   nonzero (fortran only, 1D only)
+    -   `empty`, `full`, `ones`, `zeros`, `arange` (`like` parameter is not supported).
+    -   `empty_like`, `full_like`, `zeros_like`, and `ones_like` (`subok` parameter is not supported).
+    -   `rand`, `randint`
+    -   `where`, `count_nonzero` (fortran only)
+    -   `nonzero` (fortran only, 1D only)
 
 -   others:
 
-    -   amax, amin, sum, shape, size, floor, sign
+    -   `amax`, `amin`, `sum`, `shape`, `size`, `floor`, `sign`
 
 If discrepancies beyond round-off error are found between [Numpy](https://numpy.org/doc/stable/reference/)'s and [Pyccel](https://github.com/pyccel/pyccel)'s results, please create an issue at <https://github.com/pyccel/pyccel/issues> and provide a small example of your problem. Do not forget to specify your target language.
