+++ .dict_custom.txt
@@ -72,3 +72,4 @@ docstring
 docstrings
 Codacy
 codebase
+backend
