+++ tutorial/openmp.md
@@ -65,9 +65,9 @@ Other references:
 
 [*OpenMP Clauses*](https://docs.microsoft.com/en-us/cpp/parallel/openmp/reference/openmp-clauses)
 
-### parallel Construct
+### `parallel` Construct
 
-#### Syntax of *parallel*
+#### Syntax of `parallel`
 
 ```python
 #$ omp parallel [clause[ [,] clause] ... ]
@@ -96,9 +96,9 @@ hello from thread: 0
 hello from thread: 1
 ```
 
-### loop Construct
+### `loop` Construct
 
-#### Syntax of *loop*
+#### Syntax of `loop`
 
 ```python
 #$ omp for [nowait] [clause[ [,] clause] ... ]
@@ -127,9 +127,9 @@ The output of this program is:
 893116
 ```
 
-### single Construct
+### `single` Construct
 
-#### Syntax of *single*
+#### Syntax of `single`
 
 ```python
 #$ omp single [nowait] [clause[ [,] clause] ... ]
@@ -163,9 +163,9 @@ hello from thread number:            3
 hello from thread number:            0
 ```
 
-### critical Construct
+### `critical` Construct
 
-#### Syntax of *critical*
+#### Syntax of `critical`
 
 ```python
 #$ omp critical [(name) [ [,] hint (hint-expression)]]
@@ -195,9 +195,9 @@ The output of this program is:
 893116
 ```
 
-### barrier Construct
+### `barrier` Construct
 
-#### Syntax of *barrier*
+#### Syntax of `barrier`
 
 ```python
 #$ omp barrier
@@ -233,9 +233,9 @@ The output of this program is:
 1786232
 ```
 
-### masked Construct
+### `masked` Construct
 
-#### Syntax of *masked*
+#### Syntax of `masked`
 
 ```python
 #$ omp masked [ filter(integer-expression) ]
@@ -265,14 +265,14 @@ result : 1
 
 ### `taskloop`/`atomic` Construct
 
-#### Syntax of *`taskloop`*
+#### Syntax of `taskloop`
 
 ```python
 #$ omp taskloop [clause[ [,]clause] ... ]
 for-loops
 ```
 
-#### Syntax of *atomic*
+#### Syntax of `atomic`
 
 ```python
 #$ omp atomic [clause[ [,]clause] ... ]
@@ -315,9 +315,9 @@ x1 : 200
 x2 : 100
 ```
 
-### simd Construct
+### `simd` Construct
 
-#### Syntax of *simd*
+#### Syntax of `simd`
 
 ```python
 #$ omp simd [clause[ [,]clause] ... ]
@@ -349,9 +349,9 @@ The output of this program is:
 Result: 893116
 ```
 
-### task / taskwait Construct
+### `task` / `taskwait` Construct
 
-#### Syntax of *task*
+#### Syntax of `task`
 
 ```python
 #$ omp task [clause[ [,]clause] ... ]
@@ -359,7 +359,7 @@ structured-block
 #$ omp end task
 ```
 
-#### Syntax *taskwait*
+#### Syntax `taskwait`
 
 ```python
 #$ omp taskwait
@@ -368,7 +368,7 @@ structured-block
 #### Example
 
 The ``` #$ omp task ``` pragma is used here to define an explicit task.\
-The ``` #$ omp taskwait ``` pragma is used here to specify that the current task region remains suspended until all child tasks that it generated before the taskwait construct complete execution.
+The ``` #$ omp taskwait ``` pragma is used here to specify that the current task region remains suspended until all child tasks that it generated before the `taskwait` construct complete execution.
 ```python
 @types('int', results='int')
 def fib(n):
@@ -397,9 +397,9 @@ The output of this program is:
 55
 ```
 
-### taskyield Construct
[…]
