+++ .dict_custom.txt
@@ -73,3 +73,4 @@ docstrings
 Codacy
 codebase
 backend
+iterable
