+++ pyccel/ast/c_concepts.py
@@ -10,6 +10,10 @@ Module representing object address.
 from .basic import PyccelAstNode, Basic
 from .literals  import LiteralString
 
+__all__ = ('ObjectAddress',
+           'CStringExpression',
+           'CMacro')
+
 class ObjectAddress(PyccelAstNode):
     """Represents the address of an object.
     ObjectAddress(Variable('int','a'))                            is  &a
