+++ pyccel/ast/c_concepts.py
@@ -10,9 +10,9 @@ Module representing object address.
 from .basic import PyccelAstNode, Basic
 from .literals  import LiteralString
 
-__all__ = ('ObjectAddress',
+__all__ = ('CMacro',
            'CStringExpression',
-           'CMacro')
+           'ObjectAddress')
 
 class ObjectAddress(PyccelAstNode):
     """Represents the address of an object.
