+++ pyccel/ast/scipyext.py
@@ -10,7 +10,7 @@ from numpy import pi
 from .core import Module, Import
 from .variable import Constant
 
-__all__ = ('scipy_pi_const', 'scipy_mod')
+__all__ = ('scipy_mod', 'scipy_pi_const')
 
 scipy_pi_const = Constant('float', 'pi', value=pi)
 
