+++ .github/workflows/doc_coverage.yml
@@ -1,24 +1,10 @@
-name: Pyccel tests
+name: Doc Coverage Action
 
 on:
   pull_request:
     branches: [ master ]
 
 jobs:
-  Linux:
-
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python 3.7
-        uses: actions/setup-python@v4
-        with:
-          python-version: 3.7
-      - name: Install dependencies
-        uses: ./.github/actions/linux_install
-      - name: Install python dependencies
-        uses: ./.github/actions/pip_installation
 
   DocumentationChecker:
 
@@ -44,10 +30,6 @@ jobs:
         shell: bash
       - name: Show report
         run: |
-          echo "Documentation Coverage of Head Branch"
-          echo ""
-          docstr-coverage head/pyccel 2>&1 | tail -3
-          echo ""
-          echo "Documentation Coverage of Base Branch"
-          echo ""
-          docstr-coverage base/pyccel 2>&1 | tail -3
+          docstr-coverage head/pyccel 2>head_cov
+          docstr-coverage base/pyccel 2>base_cov
+          python ci_tools/summarise_doccoverage.py head_cov base_cov $GITHUB_STEP_SUMMARY
