+++ .github/workflows/doc_coverage.yml
@@ -31,4 +31,3 @@ jobs:
         run: |
           docstr-coverage head/pyccel 2>head_cov
           docstr-coverage base/pyccel 2>base_cov
-          python ci_tools/summarise_doccoverage.py head_cov base_cov $GITHUB_STEP_SUMMARY
