+++ .github/workflows/doc_coverage.yml
@@ -29,5 +29,7 @@ jobs:
         shell: bash
       - name: Show report
         run: |
+          touch head_cov
+          touch base_cov
           docstr-coverage head/pyccel 2>head_cov
           docstr-coverage base/pyccel 2>base_cov
