+++ .github/workflows/doc_coverage.yml
@@ -29,6 +29,6 @@ jobs:
         shell: bash
       - name: Show report
         run: |
-          touch head_cov
-          touch base_cov
-          ls
+          docstr-coverage head/pyccel 2>head_cov || true
+          docstr-coverage base/pyccel 2>base_cov || true
+          python ci_tools/summarise_doccoverage.py head_cov base_cov $GITHUB_STEP_SUMMARY
