+++ ci_tools/summarise_doccoverage.py
@@ -38,13 +38,13 @@ for branch_file in [args.head, args.base]:
                 results[branch + '_no_obj'].update(objname)
             i += 1
 
-added_mod = [mod for mod in results['base_no_mod'] if mod not in results['head_no_mod']]
-added_obj = [obj for obj in results['base_no_obj'] if obj not in results['head_no_obj']]
+added_mod = [mod for mod in results['compare_no_mod'] if mod not in results['base_no_mod']]
+added_obj = [obj for obj in results['compare_no_obj'] if obj not in results['base_no_obj']]
 with open(args.output, 'w', encoding="utf-8") as f:
     print('Base Branch Summary', file=f)
     print(results['base_summary'], file=f)
-    print('Head Branch Summary', file=f)
-    print(results['head_summary'], file=f)
+    print('Compare Branch Summary', file=f)
+    print(results['compare_summary'], file=f)
     if len(added_mod) > 0:
         print('This pull request added these modules without docstrings:', file=f)
         for mod in added_mod:
