+++ .github/workflows/doc_coverage.yml
@@ -13,11 +13,11 @@ jobs:
     steps:
       - uses: actions/checkout@v3
         with:
-          ref: ${{ github.event.pull_request.head.sha }}
-          path: head
+          ref: ${{ github.event.pull_request.base.sha }}
+          path: base
       - uses: actions/checkout@v3
         with:
-          path: base
+          path: compare
       - name: Set up Python 3.7
         uses: actions/setup-python@v4
         with:
@@ -29,9 +29,7 @@ jobs:
         shell: bash
       - name: Show report
         run: |
-          cat head/pyccel/ast/basic.py
-          cat base/pyccel/ast/basic.py
-          docstr-coverage head/pyccel 2>head_cov || true
-          docstr-coverage base/pyccel 2>base_cov || true
-          python base/ci_tools/summarise_doccoverage.py head_cov base_cov $GITHUB_STEP_SUMMARY
+          docstr-coverage head/pyccel 2>base_cov || true
+          docstr-coverage base/pyccel 2>compare_cov || true
+          python compare/ci_tools/summarise_doccoverage.py head_cov base_cov $GITHUB_STEP_SUMMARY
           cat $GITHUB_STEP_SUMMARY
