### 2023-01-09 38b2d10
Add current_context to variable in _infere_type
