+++ ci_tools/summarise_doccoverage.py
@@ -29,7 +29,6 @@ for branch_file in [args.base, args.compare]:
     while i < n:
         modname = lines[i].split()[1].strip('"')[:-3].replace('/','.').split(branch, 1)[1][1:]
         i+=1
-        words = set()
         while i<n and lines[i].startswith(' - '):
             if lines[i].startswith(' - No module docstring'):
                 results[branch + '_no_mod'].update(modname)
@@ -38,6 +37,7 @@ for branch_file in [args.base, args.compare]:
                 results[branch + '_no_obj'].update(objname)
             i += 1
 
+print(results['compare_no_obj'])
 added_mod = [mod for mod in results['compare_no_mod'] if mod not in results['base_no_mod']]
 added_obj = [obj for obj in results['compare_no_obj'] if obj not in results['base_no_obj']]
 with open(args.output, 'w', encoding="utf-8") as f:
