+++ .github/workflows/doc_coverage.yml
@@ -29,7 +29,6 @@ jobs:
         shell: bash
       - name: Show report
         run: |
-          diff base/pyccel/ast/basic.py compare/pyccel/ast/basic.py
           docstr-coverage base/pyccel 2>base_cov || true
           docstr-coverage compare/pyccel 2>compare_cov || true
           python compare/ci_tools/summarise_doccoverage.py compare_cov base_cov $GITHUB_STEP_SUMMARY
