### 2023-01-09 1374b52
update tmp_strides and tmp_shape starting from an offset counted using the first slice with range greater than 1
