+++ developer_docs/order_docs.md
@@ -2,12 +2,12 @@
 
 ## Order in numpy
 
-`order` is the parameter given to the `numpy.array` function in order to choose how the array is stored in memory.
-Both of the orders discussed here (`order_c` and `order_f`) are stored **contiguously** in memory, but they differ in how they are arranged.
+`order` is the parameter given to the `numpy.array` function in order to choose how a multi-dimensional array is stored in memory.
+For both of the orders discussed here (`C` and `F`) the arrays are stored **contiguously** in memory, but they differ in how their entries are arranged.
 
 ### Order C
 
-`order='C'` tells `Numpy` to store the array row by row (row-major), example:
+`order='C'` tells Numpy to store the array row by row (row-major). For example:
 
 ```python
 import numpy as np
@@ -20,12 +20,12 @@ if __name__ == "__main__":
  ```
 
 `array.ravel('k')` shows us how the array is stored in memory.
-This python script will output `[1 2 3 4 5 6 7 8 9]`, notice that the rows are stored one after the other.
-This is the default behaviour in python.
+This Python script will output `[1 2 3 4 5 6 7 8 9]`; notice that the rows are stored one after the other.
+This is the default behaviour in Python.
 
 ### Order F
 
-`order='F'` on the other hand tells `Numpy` to store the array column by column (column-major), example:
+`order='F'` on the other hand tells Numpy to store the array column by column (column-major). For example:
 
 ```python
 import numpy as np
@@ -37,11 +37,11 @@ if __name__ == "__main__":
   print(a.ravel('K'))
 ```
 
-This python script will output `[1 4 7 2 5 8 3 6 9]`, notice that the columns are stored one after the other.
+This Python script will output `[1 4 7 2 5 8 3 6 9]`; notice that the columns are stored one after the other.
 
-### Printing and indexing in `Numpy`
+### Printing and indexing in Numpy
 
-`order` in `Numpy` does not affect the indexing or the printing, unlike `transposing`, the `shape` of the array remains the same, only the `strides` change, example:
+The `order` of a Numpy array does not affect the indexing or the printing: unlike the `transpose` operation, the `shape` of the array remains the same, and only the `strides` change. For example:
 
 ```python
 import numpy as np
@@ -68,7 +68,7 @@ if __name__ == "__main__":
    print(b[2][1], b[0][0], b[1]) # output: 8 1 [4 5]
 ```
 
-`arr.strides` is the variable that helps us navigate the array (indexing/printing), it tells us how many bytes we have to skip in memory to move to the next position along a certain axis (dimension). For example for `memory_layout_of_a = [1 4 7 2 5 8]` and `strides_of_a = (8, 24)`, we have to skip 8 bytes (1 element for `int64`) to move to the next row, but 24 bytes (3 elements for `int64`) to get to the same position in the next column of `a`.
+`arr.strides` is the variable that helps us navigate the array (indexing/printing) by telling us how many bytes we have to skip in memory to move to the next position along a certain axis (dimension). For example for `memory_layout_of_a = [1 4 7 2 5 8]` and `strides_of_a = (8, 24)`, we have to skip 8 bytes (1 element for `int64`) to move to the next row, but 24 bytes (3 elements for `int64`) to get to the same position in the next column of `a`.
 `a[2][1]` would give us `'8'`, using the `strides`: `2 * 8 + 1 * 24 = 40`, which means that in the flattened array, we would have to skip `40` bytes to get the value of `a[2][1]`, each element is 8 bytes, so we would have to skip `40 / 8 = 5` elements to get to `'8'`
 
 The order does however change how the user writes code.
@@ -95,11 +95,11 @@ if __name__ == "__main__":
 
[…]
