+++ ci_tools/summarise_pyspelling.py
@@ -38,23 +38,24 @@ while i < n:
 
 if errors:
     all_words = set()
+
+    with open(os.path.join(os.path.dirname(__file__),'..','.dict_custom.txt'), encoding="utf-8") as d:
+        internal_dict = [w.strip() for w in d.readlines()]
+
     with open(args.output, 'w', encoding="utf-8") as f:
         print("There are misspelled words", file=f)
         for name, words in errors.items():
             print("## `", name, "`", file=f)
             for w in words:
-                suggestions = difflib.get_close_matches(word, internal_dict)
+                suggestions = difflib.get_close_matches(w, internal_dict)
                 if suggestions:
-                    print("-   ", w, "  f:  Did you mean {word} -> {suggestions}", file=f)
+                    print("-   ", w, "  f:  Did you mean {w} -> {suggestions}", file=f)
                 else:
                     print("-   ", w, file=f)
             print(file=f)
             all_words.update(words)
 
         print("These errors may be due to typos, capitalisation errors, or lack of quotes around code. If this is a false positive please add your word to `.dict_custom.txt`", file=f)
-
-        with open(os.path.join(os.path.dirname(__file__),'..','.dict_custom.txt'), encoding="utf-8") as d:
-            internal_dict = [w.strip() for w in d.readlines()]
 
     sys.exit(1)
 else:
