### 2023-01-09 459630e
Fix return Constants bug (#1283)
Ensure results variables exist locally (fixes #1282). Add unit test
