### 2023-01-09 8b72efb
mention the dimensionality of array in the kernel example.
Co-authored-by: Fatima-zahra RAMDANI <52450718+framdani@users.noreply.github.com>
