### 2023-01-09 1c5307e
check dimension
