### 2023-01-09 f33701b
remove unecessary error check
