### 2023-01-09 261c1d9
Improve CI (#1298)
- Split up CI to allow for better summaries
- Add a CI to test spelling
- Clean up spelling in markdown files
- Add a CI to ensure Pyccel conventions are followed (`__slots__`, `__all__` and `_attribute_nodes` in `pyccel.ast`)
- Fix classes which didn't correctly follow conventions.
