+++ README.md
@@ -16,10 +16,10 @@ The aim of **Pyccel** is to provide a simple way to generate automatically, para
 
 Pyccel comes with a selection of **extensions** allowing you to convert calls to some specific Python packages to Fortran/C. The following packages will be covered (partially):
 
--   numpy
--   scipy
--   mpi4py (not available yet)
--   h5py (not available yet)
+-   `numpy`
+-   `scipy`
+-   `mpi4py` (not available yet)
+-   `h5py` (not available yet)
 
 Pyccel's acceleration capabilities lead to much faster code. Comparisons of Python vs Pyccel or other tools can be found in the [benchmarks](https://github.com/pyccel/pyccel-benchmarks) repository.
 The results for the master branch currently show the following performance on python 3.10:
@@ -42,7 +42,7 @@ If you are eager to try Pyccel out, we recommend reading our [quick-start guide]
     -   [Windows](#Windows)
 
 -   [Installation](#Installation)
-    -   [From PyPi](#From-PyPi)
+    -   [From PyPI](#From-PyPI)
     -   [From sources](#From-sources)
     -   [On a read-only system](#On-a-read-only-system)
 
@@ -72,7 +72,7 @@ If you are eager to try Pyccel out, we recommend reading our [quick-start guide]
 
 -   Supported libraries/APIs
     -   [OpenMP](./tutorial/openmp.md)
-    -   [Numpy](./tutorial/numpy-functions.md)
+    -   [NumPy](./tutorial/numpy-functions.md)
 
 ## Pyccel Installation Methods
 
@@ -82,8 +82,8 @@ Some advanced features of Pyccel require additional non-Python libraries to be i
 Alternatively, Pyccel can be deployed through a **Linux Docker image** that contains all dependencies, and which can be setup with any version of Pyccel.
 For more information, please read the section on [Pyccel container images](#Pyccel-Container-Images).
 
-It is possible to use pyccel with anaconda but this is generally not advised as anaconda modifies paths used for finding executables, shared libraries and other objects.
-Support is provided for anaconda on linux/macos.
+It is possible to use Pyccel with anaconda but this is generally not advised as anaconda modifies paths used for finding executables, shared libraries and other objects.
+Support is provided for anaconda on linux/macOS.
 
 On Windows support is limited to examples which do not use external libraries.
 This is because we do not know of a way to reliably avoid [DLL hell](https://en.wikipedia.org/wiki/DLL_Hell).
@@ -208,7 +208,7 @@ msiexec //i msmpisdk.msi
 
 At this point, close and reopen your terminal to refresh all environment variables!
 
-In Administrator git-bash, generate mpi.mod for gfortran according to <https://abhilashreddy.com/writing/3/mpi_instructions.html>:
+In Administrator git-bash, generate `mpi.mod` for GFortran according to <https://abhilashreddy.com/writing/3/mpi_instructions.html>:
 
 ```sh
 cd "$MSMPI_INC"
@@ -218,7 +218,7 @@ gfortran -c -D_WIN64 -D INT_PTR_KIND\(\)=8 -fno-range-check mpi.f90
 cd -
 ```
 
-Generate static libmsmpi.a from msmpi.dll:
+Generate static `libmsmpi.a` from `msmpi.dll`:
 
 ```sh
 cd "$MSMPI_LIB64"
@@ -254,7 +254,7 @@ echo "import os; os.add_dll_directory('C://ProgramData/chocolatey/lib/mingw/tool
 
 On Windows and/or Anaconda Python, use `pip` instead of `pip3` for the Installation of Pyccel below.
 
-### From PyPi
+### From PyPI
 
 Simply run, for a user-specific installation:
 
@@ -288,7 +288,7 @@ for a system-wide installation.
     pip3 install --user -e .[test]
     ```
 
[…]
