+++ tutorial/decorators.md
@@ -117,7 +117,7 @@ positive or negative. As a result an if block must be added. This implies a (pot
 cost. Non-literal negative indexes are not especially common, therefore Pyccel does not add this costly
 if block unless it is specifically requested. This can be done using the `allow_negative_index` decorator.
 
-An example shows how Pyccel handles negative indexes beween Python and C:
+An example shows how Pyccel handles negative indexes between Python and C:
 
 ```python
 from pyccel.decorators import allow_negative_index
@@ -209,7 +209,7 @@ end module boo
 
 ## Elemental
 
-In Python it is often the case that a function with scalar arguments and a single scalar output (if any) is also able to accept Numpy arrays with identical rank and shape - in such a case the scalar function is simply applied element-wise to the input arrays. In order to mimic this behavior in the generated C or Fortran code, Pyccel provides the decorator `elemental`.
+In Python it is often the case that a function with scalar arguments and a single scalar output (if any) is also able to accept NumPy arrays with identical rank and shape - in such a case the scalar function is simply applied element-wise to the input arrays. In order to mimic this behaviour in the generated C or Fortran code, Pyccel provides the decorator `elemental`.
 
 Important note: applying the `elemental` decorator to a function will not make a difference to the C translation of the function definition itself since C doesn't have the elementwise feature. However, Pyccel implements that functionality by calling the function in a `for` loop when an array argument is passed. In the following example, we will use the function `square` where `@elemental` will be useful:
 
@@ -602,9 +602,9 @@ The generated C code:
 
 ## Getting Help
 
-If you face problems with pyccel, please take the following steps:
+If you face problems with Pyccel, please take the following steps:
 
-1.  Consult our documention in the tutorial directory;
+1.  Consult our documentation in the tutorial directory;
 2.  Send an email message to pyccel@googlegroups.com;
 3.  Open an issue on GitHub.
 
