+++ developer_docs/semantic_stage.md
@@ -4,7 +4,7 @@ The semantic stage is described by the file [pyccel.parser.semantic](../pyccel/p
 
 The semantic stage serves several purposes:
 1.  [**Types**](#Types) : Determine the type of each symbol
-2.  [**Impose Restrictions**](#Impose-restrictions) : Ensure that the code follows any restictions that Pyccel imposes
+2.  [**Impose Restrictions**](#Impose-restrictions) : Ensure that the code follows any restrictions that Pyccel imposes
 3.  [**Function Recognition**](#Function-recognition) : Identify any functions that are recognised by Pyccel
 4.  [**Imports**](#Imports) : Identify imports
 5.  [**Low-level Objects**](#Low-level-objects) : Create any objects which are hidden in high-level code, but must appear explicitly in low-level code
@@ -32,7 +32,7 @@ Variables and objects which can be saved in variables (e.g. literals and arrays)
 The type indicates all the information that allows the object to be declared in a low-level language.
 The interface to access these characteristics is defined in the super class [`pyccel.ast.basic.PyccelAstNode`](../pyccel/ast/basic.py).
 The characteristics are:
--   **data type** : bool/int/float/complex/class type/etc
+-   **data type** : boolean/integer/float/complex/class type/etc
 -   **precision** : The number of bytes required to store an object of this data type
 -   **rank** : The number of dimensions of the array (0 for a scalar)
 -   **shape** : The number of elements in each dimension of the array (`()` for a scalar)
@@ -84,15 +84,15 @@ This function takes several arguments (see docstring for more details).
 The most important arguments are:
 -   _message_ : Describe the issue that lead to the error
 
--   _symbol_ : The Python ast object should be passed here. This object contains information about its position in the file (line number, column) which ensures the user can more easily locate their error
+-   _symbol_ : The Python AST object should be passed here. This object contains information about its position in the file (line number, column) which ensures the user can more easily locate their error
 
 -   _severity_ : The severity level must be one of the following:
     -   _warning_ : An error will be printed but Pyccel will continue executing
     -   _error_ : An error will be printed but Pyccel will continue executing the syntactic stage
-    -   _fatal_ : An error will be printed and Pyccel will stop executing. This is the level used the most frequently in the semantic stage as, if information such as the type cannot be determined, the object cannot be created which will cause problems later in the exection.
+    -   _fatal_ : An error will be printed and Pyccel will stop executing. This is the level used the most frequently in the semantic stage as, if information such as the type cannot be determined, the object cannot be created which will cause problems later in the execution.
 
 It should be noted that Pyccel assumes that the user has provided valid code.
-It is not feasable to provide error messages for every possible coding error so we limit ourselves to simple validity checks and Pyccel restriction errors.
+It is not feasible to provide error messages for every possible coding error so we limit ourselves to simple validity checks and Pyccel restriction errors.
 
 ## Function Recognition
 
@@ -103,7 +103,7 @@ In this case the `funcdef` attribute of the [`pyccel.ast.core.FunctionCall`](../
 The type of the `FunctionCall` (required to assign the result to a `Variable`) is easily determined from the type of the result variable(s).
 
 The second case involves functions that are recognised by Pyccel.
-This includes built-in functions and functions imported from supported libraries (numpy, math, etc).
+This includes built-in functions and functions imported from supported libraries (`numpy`, `math`, etc).
 Built-in functions do not need importing.
 Instead they are recognised via the function [`pyccel.ast.utilities.builtin_function`](../pyccel/ast/utilities.py) which uses the dictionary [`pyccel.ast.builtins.builtin_functions_dict`](../pyccel/ast/builtins.py) to identify supported functions.
[…]
