+++ .dict_custom.txt
@@ -74,3 +74,5 @@ Codacy
 codebase
 backend
 iterable
+performant
+inhomogeneous
