### 2023-01-10 6d8489d
Add case of PyccelArraySize for cupy ravel
