+++ tests/internal/scripts/ccuda/cupy_arange.py
@@ -12,6 +12,6 @@ if __name__ == '__main__':
     threads_per_block = 32
     n_blocks = 1
     arr = cp.arange(32)
-    cuda.deviceSynchronize()
+    cuda.synchronize()
     func[n_blocks, threads_per_block](arr)
-    cuda.deviceSynchronize()
+    cuda.synchronize()
