### 2023-01-10 d0854c4
Put memory_location assignation in CupyNewArray
