### 2023-01-10 0031f24
Fix codacy errors
