### 2023-01-10 78cc7b8
removing output file
