### 2023-01-10 19a7ad2
Remove memory_location from cupy parameter / add memory_location property
