+++ pyccel/ast/builtins.py
@@ -1,951 +1,0 @@
-# coding: utf-8
-#------------------------------------------------------------------------------------------#
-# This file is part of Pyccel which is released under MIT License. See the LICENSE file or #
-# go to https://github.com/pyccel/pyccel/blob/master/LICENSE for full license details.     #
-#------------------------------------------------------------------------------------------#
-"""
-The Python interpreter has a number of built-in functions and types that are
-always available.
-
-In this module we implement some of them in alphabetical order.
-
-"""
-from pyccel.errors.errors import PyccelError
-
-from pyccel.utilities.stage import PyccelStage
-
-from .basic     import Basic, PyccelAstNode
-from .datatypes import (NativeInteger, NativeBool, NativeFloat,
-                        NativeComplex, NativeString, str_dtype,
-                        NativeGeneric, default_precision)
-from .internals import PyccelInternalFunction, max_precision, Slice
-from .literals  import LiteralInteger, LiteralFloat, LiteralComplex, Nil
-from .literals  import Literal, LiteralImaginaryUnit, get_default_literal_value
-from .literals  import LiteralString
-from .operators import PyccelAdd, PyccelAnd, PyccelMul, PyccelIsNot
-from .operators import PyccelMinus, PyccelUnarySub, PyccelNot
-from .variable  import IndexedElement
-
-pyccel_stage = PyccelStage()
-
-__all__ = (
-    'Lambda',
-    'PythonAbs',
-    'PythonComplexProperty',
-    'PythonReal',
-    'PythonImag',
-    'PythonConjugate',
-    'PythonBool',
-    'PythonComplex',
-    'PythonEnumerate',
-    'PythonFloat',
-    'PythonInt',
-    'PythonTuple',
-    'PythonLen',
-    'PythonList',
-    'PythonMap',
-    'PythonPrint',
-    'PythonRange',
-    'PythonSum',
-    'PythonType',
-    'PythonZip',
-    'PythonMax',
-    'PythonMin',
-    'python_builtin_datatype'
-)
-
-#==============================================================================
-class PythonComplexProperty(PyccelInternalFunction):
-    """Represents a call to the .real or .imag property
-
-    e.g:
-    > a = 1+2j
-    > a.real
-    1.0
-
-    arg : Variable, Literal
-    """
-    __slots__ = ()
-    _dtype = NativeFloat()
-    _precision = -1
-    _rank  = 0
-    _shape = None
-    _order = None
-
-    def __init__(self, arg):
-        super().__init__(arg)
-
-    @property
-    def internal_var(self):
-        """Return the variable on which the function was called"""
-        return self._args[0]
-
-#==============================================================================
-class PythonReal(PythonComplexProperty):
-    """Represents a call to the .real property
-
-    e.g:
-    > a = 1+2j
-    > a.real
-    1.0
-
-    arg : Variable, Literal
-    """
-    __slots__ = ()
-    name = 'real'
-    def __new__(cls, arg):
-        if isinstance(arg.dtype, NativeBool):
-            return PythonInt(arg)
-        elif not isinstance(arg.dtype, NativeComplex):
-            return arg
-        else:
-            return super().__new__(cls)
-
-    def __str__(self):
-        return 'Real({0})'.format(str(self.internal_var))
-
-#==============================================================================
-class PythonImag(PythonComplexProperty):
-    """Represents a call to the .imag property
-
-    e.g:
-    > a = 1+2j
-    > a.imag
-    1.0
-
-    arg : Variable, Literal
-    """
-    __slots__ = ()
-    name = 'imag'
-    def __new__(cls, arg):
-        if arg.dtype is not NativeComplex():
-            return get_default_literal_value(arg.dtype)
-        else:
-            return super().__new__(cls)
-
[…]
