### 2023-01-10 397887b
printing results
