+++ .github/workflows/doc_coverage.yml
@@ -36,6 +36,6 @@ jobs:
           BASE_BRANCH=$GITHUB_BASE_REF
           git fetch
           git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color
-          cat pull_diff.txt
           python ci_tools/list_changed_files.py pull_diff.txt changed_files.txt
+          cat changed_files.txt
 
