### 2023-01-10 4a8fab5
Put current context check in _assign_lhs_variable
