+++ .github/workflows/doc_coverage.yml
@@ -35,7 +35,7 @@ jobs:
           cd compare
           BASE_BRANCH=$GITHUB_BASE_REF
           git fetch
-          git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color
+          git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color &>/dev/null
           python ci_tools/list_changed_files.py pull_diff.txt changed_files.txt
-          cat changed_files.txt
-
+          python ci_tools/list_docs_tovalidate.py changed_files.txt objects.txt
+          cat objects.txt
