+++ .github/workflows/doc_coverage.yml
@@ -34,8 +34,9 @@ jobs:
           python compare/ci_tools/summarise_doccoverage.py compare_cov base_cov
           cd compare
           BASE_BRANCH=$GITHUB_BASE_REF
-          git fetch
+          git fetch &>/dev/null
           git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color &>/dev/null
           python ci_tools/list_changed_files.py pull_diff.txt changed_files.txt
+          cat changed_files.txt
           python ci_tools/list_docs_tovalidate.py changed_files.txt objects.txt
           cat objects.txt
