### 2023-01-10 8322ec1
add 3d tests
