### 2023-01-10 e8db6bc
Add memory_location to infere_type for CupyNewArray
