### 2023-01-10 bd6faa6
Clean up cupy_arrange and cupy_array tests
