### 2023-01-10 d128a37
running numpydoc
