+++ .github/workflows/doc_coverage.yml
@@ -26,7 +26,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install docstr-coverage
-          pythom -m pip install numpydoc
+          python -m pip install numpydoc
         shell: bash
       - name: Show report
         run: |
