+++ .github/workflows/doc_coverage.yml
@@ -38,11 +38,9 @@ jobs:
           git fetch &>/dev/null
           git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color &>/dev/null
           python ci_tools/list_changed_files.py pull_diff.txt changed_files.txt
-          cat changed_files.txt
           python ci_tools/list_docs_tovalidate.py changed_files.txt objects.txt
-          cat objects.txt
           while read line; do    
             echo $line >> report.txt
-            python3 -m numpydoc --validate $line >> report.txt || true 
+            python3 -m numpydoc --validate $line >> report.txt 2>&1 || true 
           done < objects.txt
           diff objects.txt report.txt && echo "SUCCESS! All docstrings are valid" || echo "FAILURE" && cat report.txt && exit 1
