+++ .github/workflows/doc_coverage.yml
@@ -38,6 +38,7 @@ jobs:
           git fetch &>/dev/null
           git diff origin/${BASE_BRANCH}..HEAD --no-indent-heuristic --unified=0 --output=pull_diff.txt --no-color &>/dev/null
           python ci_tools/list_changed_files.py pull_diff.txt changed_files.txt
+          touch objects.txt
           python ci_tools/list_docs_tovalidate.py changed_files.txt objects.txt
           while read line; do    
             echo $line >> report.txt
