+++ .github/workflows/doc_coverage.yml
@@ -45,4 +45,4 @@ jobs:
             echo $line >> report.txt
             python3 -m numpydoc --validate $line >> report.txt 2>&1 || true 
           done < objects.txt
-          diff objects.txt report.txt && echo "SUCCESS! All docstrings are valid" || echo "FAILURE" && cat report.txt && exit 1
+          (diff objects.txt report.txt && echo "SUCCESS! All docstrings are valid") || (echo "FAILURE" && cat report.txt && exit 1)
