### 2023-01-10 c159d7f
Update kernel decorator definition
