### 2023-01-10 8ef5990
Revert cuda_array_managed.py to cuda_main_temp version
