### 2023-01-10 b42d117
cleaned cuda tests to be more specific
