### 2023-01-10 4691db4
Fix CudaSynchronize printer's name
