### 2023-01-10 5f32e4f
Add test containing only a call to cuda.synchronize
