+++ tests/internal/scripts/ccuda/cuda_synchronize.py
@@ -1,0 +1,5 @@
+from pyccel import cuda
+
+if __name__ == '__main__':
+    cuda.synchronize()
+    cuda.synchronize()
