### 2023-01-10 93c270c
fixed kernel_launch function
