### 2023-01-10 83a93f5
simplified view tests
