### 2023-01-10 2a851ed
Use only one cuda.synchronize in cuda_synchronize.py
