+++ tests/internal/scripts/ccuda/cuda_synchronize.py
@@ -2,4 +2,3 @@ from pyccel import cuda
 
 if __name__ == '__main__':
     cuda.synchronize()
-    cuda.synchronize()
