### 2023-01-10 17180eb
Make memory location a class attribute for CupyNewArray
