### 2023-01-11 a8c5db4
Clean up cupy ravel tests
