### 2023-01-11 631fd9b
Fix type inference for Cupy array functions (#1300)
* Inital issue fix

* Put memory_location assignation in CupyNewArray

* Fix codacy errors

* Remove memory_location from cupy parameter / add memory_location property

* Fix the slots

* Add memory_location to infere_type for CupyNewArray

* Clean up cupy_arrange and cupy_array tests

* Make memory location a class attribute for CupyNewArray
