+++ pyccel/ast/cupyext.py
@@ -43,6 +43,7 @@ errors = Errors()
 pyccel_stage = PyccelStage()
 
 __all__ = (
+    'CupyNewArray',
     'CupyArray',
     'CupyEmpty',
     'CupyEmptyLike',
@@ -58,7 +59,22 @@ __all__ = (
 )
 
 #==============================================================================
-class CupyArray(CudaNewArray):
+class CupyNewArray(CudaNewArray):
+    """ Class from which all Cupy functions which imply a call to Allocate
+    inherit
+    """
+    _memory_location = 'device'
+    def __init__(self):
+        super().__init__()
+
+    @property
+    def memory_location(self):
+        """ Indicate if the array is allocated on the host, device or has a managed memory
+        """
+        return self._memory_location
+
+#==============================================================================
+class CupyArray(CupyNewArray):
     """
     Represents a call to  cupy.array for code generation.
 
@@ -123,7 +139,7 @@ class CupyArray(CudaNewArray):
         return self._arg
 
 #==============================================================================
-class CupyArange(CudaNewArray):
+class CupyArange(CupyNewArray):
     """
     Represents a call to  cupy.arange for code generation.
 
@@ -198,7 +214,7 @@ class Shape(PyccelInternalFunction):
             return PythonTuple(*arg.shape)
 
 #==============================================================================
-class CupyFull(CudaNewArray):
+class CupyFull(CupyNewArray):
     """
     Represents a call to cupy.full for code generation.
 
@@ -242,7 +258,7 @@ class CupyFull(CudaNewArray):
         self._shape = shape
         self._rank  = len(self._shape)
         self._dtype = dtype
-        self._order = CudaNewArray._process_order(self._rank, order)
+        self._order = CupyNewArray._process_order(self._rank, order)
         self._precision = precision
 
         super().__init__(fill_value)
