+++ developer_docs/overview.md
@@ -10,7 +10,7 @@ Pyccel's development is split into 4 main stages:
 -   [Semantic Stage](#Semantic-stage)
 -   [Code Generation Stage](#Code-generation-stage)
 -   [Wrapping Stage](#Wrapping-stage)
--   [Compliation Stage](#Compilation-stage)
+-   [Compilation Stage](#Compilation-stage)
 
 ### Syntactic Stage
 
@@ -26,7 +26,7 @@ The role of this stage has decreased significantly since we moved from [redbaron
 
 This is the most important stage in Pyccel. It is here that all the information about types is calculated. This stage strives to be **language-agnostic**; this means for example, that additional variables required to handle problems appearing in one specific language should not be created here.
 
-When adding functions to this stage the aim is often to create a `PyccelAstNode` (see [ast/basic.py](../Pyccel/ast/basic.py)) and correctly define all of its parameters. This information is sometimes readily available (e.g. the type of a `PyccelAdd` can be derived from the type of the variables passed to it), but sometimes the information must be collected from elsewhere (e.g. when creating a `Variable` from a `PyccelSymbol` (roughly equivalent to a string). In this case information is needed from a `Scope` instance which is stored in the `scope`.
+When adding functions to this stage the aim is often to create a `PyccelAstNode` (see [ast/basic.py](../Pyccel/ast/basic.py)) and correctly define all of its parameters. This information is sometimes readily available (e.g. the type of a `PyccelAdd` can be derived from the type of the variables passed to it), but sometimes the information must be collected from elsewhere (e.g. when creating a `Variable` from a `PyccelSymbol`, which is roughly equivalent to a string). In this case information is needed from a `Scope` instance which is stored in the `scope`.
 
 In computer science, the _scope_ is the area of a program where an item (e.g. variable, function, etc.) is recognised. For example a variable defined in a function will not be recognised outside of that function, therefore the function defines its scope.
 
@@ -40,7 +40,7 @@ As in the Semantic stage, the Code Generation stage also stores the current Scop
 
 ### Wrapping Stage
 
-In this stage, Pyccel creates a wrapper to interface the generated low-level code with python.
+In this stage, Pyccel creates a wrapper to interface the generated low-level code with Python. The wrapper is a C file which uses the [Python/C API](https://docs.python.org/3/c-api/intro.html) and the [NumPy C-API](https://numpy.org/doc/stable/reference/c-api/index.html).
 
 ### Compilation Stage
 
@@ -66,7 +66,7 @@ Therefore the print functions which are acceptable for visiting this object are:
 
 We run through these possible functions choosing the one which is the most specialised. If none of these methods exist, then an error is raised.
 
-In the case of `NumpyTanh` the function which will be selected is `_print_NumpyUfuncBase` when translating to C or Fortran, and `_print_PyccelInternalFunction` when translating to Python
+In the case of `NumpyTanh` the function which will be selected is `_print_NumpyUfuncBase` when translating to C or Fortran, and `_print_PyccelInternalFunction` when translating to Python.
 
 ### AST
 
@@ -86,4 +86,4 @@ If the error prevents further translation (e.g. the type of an object is now unk
 
 ## Getting Help
 
-While discussions within the associated GitHub issue are often sufficient, should you require more help do not hesitate to ask one of the other developers to add you to our slack: <pyccel.slack.com>
+While discussions within the associated GitHub issue are often sufficient, should you require more help do not hesitate to ask one of the other developers to add you to our Slack channel <pyccel.slack.com>.
