### 2023-01-11 419cb54
Small improvements to syntactic_stage.md
