### 2023-01-11 81bd653
Turn on tests for cuda
