### 2023-01-11 ef45c74
Turn on tests for cuda
