+++ .github/workflows/lint.yml
@@ -1,7 +1,7 @@
 name: Python Linting
 on:
   pull_request:
-    branches: [ master ]
+    branches: [ master, cuda_main_temp, cuda_main ]
 
 jobs:
   Linter:
