### 2023-01-11 30750a3
Fix kernel launch configuration errors (#1279)
* fix kernel launch configuration errors

* fix linter

* Avoid checking for attribute 'decorators' for Interface objects

* update error messages

* add tests

* Fix new line missing

* add missing newline

* Combining error tests in a single file

* remove Trailing whitespace

* add test for invalid kernel configuration

* fix codacy issue

* trailing whitespace removed

* deactivate numpy's recent version temp

* fix typo

* Handle 'No new line at end of file'

* Fix line definition

* Add line break between errors

* Correct url

* add tests


[…]
