### 2023-01-11 d63e27f
Clarify the need for EmptyNode
