+++ pyccel/ast/core.py
@@ -57,9 +57,9 @@ __all__ = (
     'Declare',
     'Decorator',
     'Del',
+    'DoConcurrent',
     'DottedFunctionCall',
     'Duplicate',
-    'DoConcurrent',
     'EmptyNode',
     'ErrorExit',
     'Exit',
@@ -74,10 +74,11 @@ __all__ = (
     'If',
     'IfSection',
     'Import',
-    'InlineFunctionDef',
     'InProgram',
+    'InlineFunctionDef',
     'Interface',
     'Iterable',
+    'KernelCall',
     'Module',
     'ModuleHeader',
     'Pass',
@@ -92,8 +93,8 @@ __all__ = (
     'SympyFunction',
     'While',
     'With',
-    'create_variable',
     'create_incremented_string',
+    'create_variable',
     'get_iterable_ranges',
     'inline',
     'subs'
