### 2023-01-11 68a3b11
cleaning compile stage Cuda tests. (#1301)
* cleaned cuda tests to be more specific

* fixed kernel_launch function

* added kernel_launch_config tests

* fixing pyccel linter
