+++ tests/internal/scripts/ccuda/cuda_array_host.py
@@ -1,12 +1,4 @@
-from pyccel.decorators import kernel, types
 from pyccel import cuda
-
-@types('int[:]')
-def multiplyPrintElements(a):
-    for i in a:
-        i = i * 2
-        print(i)
 
 if __name__ == '__main__':
      arr = cuda.array([0,1,2,3,4], memory_location='host')
-     multiplyPrintElements(arr)
