+++ tests/internal/scripts/ccuda/cuda_grid.py
@@ -18,15 +18,3 @@ def func_2d(a):
 def func_3d(a):
     i, j, k = cuda.grid(2)
     print("3 dim :", a[i], a[j], a[k])
-
-if __name__ == '__main__':
-    threads_per_block = 5
-    n_blocks = 1
-    arr = cuda.array([0, 1, 2, 3, 4])
-    cuda.synchronize()
-    func_1d[n_blocks, threads_per_block](arr)
-    # Since we dont support multi-dim n_block / threads_per_block
-    # func_2d and func_3d won't compile
-    # func_2d[n_blocks, threads_per_block](arr)
-    # func_3d[n_blocks, threads_per_block](arr)
-    cuda.synchronize()
