+++ tests/internal/scripts/ccuda/kernel_launch_config_literal.py
@@ -1,0 +1,11 @@
+from pyccel.decorators import kernel
+from pyccel import cuda
+
+
+@kernel
+def func():
+    i = cuda.threadIdx(0) + cuda.blockIdx(0) * cuda.blockDim(0)
+    print("Hello World! ")
+
+if __name__ == '__main__':
+    func[1, 5]()
