+++ .github/workflows/master.yml
@@ -2,7 +2,7 @@ name: master_tests
 
 on:
   push:
-    branches: [ master ]
+    branches: [ master, cuda_main ]
 
 jobs:
 
