### 2023-01-11 2d46f69
Add cuda tests
