+++ .github/workflows/Github_pytest.yml
@@ -91,7 +91,7 @@ jobs:
 
   Cuda:
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
     container: nvidia/cuda:11.7.1-devel-ubuntu20.04
     if: github.event.pull_request.base.ref != 'master'
     steps:
