+++ ci_tools/list_changed_files.py
@@ -10,7 +10,6 @@ parser.add_argument('gitdiff', metavar='gitdiff', type=str,
 parser.add_argument('result', metavar='result', type=str,
                     help='File to save the results')
 args = parser.parse_args()
-<<<<<<< HEAD
 results = gdj(args.gitdiff)
 with open(args.result,'w', encoding='utf-8') as out:
     for file,changes in results:
@@ -18,17 +17,3 @@ with open(args.result,'w', encoding='utf-8') as out:
             print(file, line, sep=' ', file=out)
         for line in changes['deletion']:
             print(file, line, sep=' ', file=out)
-=======
-
-with open(args.gitdiff, encoding="utf-8") as f:
-    lines = f.readlines()
-changes = []
-for idx, l in enumerate(lines):
-    if l.startswith('diff --git a/pyccel') and \
-        not lines[idx + 1].startswith('deleted file '):
-        changes.append(l)
-changes = [l.split()[3][2:] for l in changes]
-with open(args.result, 'w', encoding="utf-8") as f:
-    for l in changes:
-        print(l, file=f)
->>>>>>> 86324854d98bd95094345b2221bb1cddad675d59
