### 2023-01-12 pyccel/pyccel#1295 @EmilyBourne
Hello, just to give an update. The benchmarks have been tested without `-ffast-math` and the results have been reproduced (see https://github.com/pyccel/pyccel-benchmarks/pull/17).

Regarding the experimental setup. We used the GitHub runners to generate our results, however GitHub doesn't provide details of which CPUs are used. We are therefore trying to work out what machines we have access to which have a recent version of python installed. Once we have found a machine where we can run the tests we will update the paper.
