+++ ci_tools/list_changed_files.py
@@ -12,7 +12,7 @@ parser.add_argument('result', metavar='result', type=str,
 args = parser.parse_args()
 results = gdj(args.gitdiff)
 with open(args.result,'w', encoding='utf-8') as out:
-    for file,changes in results:
+    for file,changes in results.items():
         for line in changes['addition']:
             print(file, line, sep=' ', file=out)
         for line in changes['deletion']:
