+++ .github/workflows/doc_coverage.yml
@@ -18,10 +18,10 @@ jobs:
       - uses: actions/checkout@v3
         with:
           path: compare
-      - name: Set up Python 3.7
+      - name: Set up Python 3.8
         uses: actions/setup-python@v4
         with:
-          python-version: 3.7
+          python-version: 3.8
       - name: Install python dependencies
         run: |
           python -m pip install --upgrade pip
