### 2023-01-12 pyccel/pyccel#1295 @boegel
Good to hear that not using `-ffast-math` doesn't skew the results too much.

Were all benchmarks run in a single workflow run using GitHub Actions?
If not, different runs maybe end up on different types of CPUs, so that doesn't seem a reliable way of running benchmarks to me...
