### 2023-01-12 f6ee2bb
Ensure master tests are triggered on cuda_main (#1305)
Update branch with commits from master.

`cuda_main` and `cuda_devel` branches should have the same tests as
the master branch. These tests are activated here.

The setup for cuda tests is also added

Co-authored-by: Abdechahid Ihya <50792865+aihya@users.noreply.github.com>
Co-authored-by: Aymane Benaissa <47903494+Pinkyboi@users.noreply.github.com>
Co-authored-by: Youness Farini <yfarini@student.1337.ma>
Co-authored-by: Yassine Alaoui <48657685+yassine-alaoui@users.noreply.github.com>
Co-authored-by: Youness Farini <farini.work@gmail.com>
[…]
