+++ pyccel/ast/numpyext.py
@@ -32,11 +32,11 @@ from .datatypes      import (dtype_and_precision_registry as dtype_registry,
 from .internals      import PyccelInternalFunction, Slice, max_precision, get_final_precision
 from .internals      import PyccelArraySize
 
-from .literals       import LiteralInteger, LiteralFloat, LiteralComplex, convert_to_literal
+from .literals       import LiteralInteger, LiteralFloat, LiteralComplex, LiteralString, convert_to_literal
 from .literals       import LiteralTrue, LiteralFalse
 from .literals       import Nil
 from .mathext        import MathCeil
-from .operators      import broadcast, PyccelMinus, PyccelDiv
+from .operators      import broadcast, PyccelMinus, PyccelDiv, PyccelMul, PyccelAdd
 from .variable       import (Variable, Constant, HomogeneousTupleVariable)
 
 errors = Errors()
@@ -45,8 +45,14 @@ pyccel_stage = PyccelStage()
 __all__ = (
     'process_shape',
     # ---
+    'NumpyAutoFill',
+    'NumpyUfuncBase',
+    'NumpyUfuncBinary',
+    'NumpyUfuncUnary',
+    # ---
     'NumpyAbs',
     'NumpyFloor',
+    'NumpySign',
     # ---
     'NumpySqrt',
     'NumpySin',
@@ -65,18 +71,27 @@ __all__ = (
     'NumpyArccosh',
     'NumpyArctanh',
     # ---
-    'NumpyEmpty',
-    'NumpyEmptyLike',
-    'NumpyFloat',
+    'NumpyAmax',
+    'NumpyAmin',
+    'NumpyArange',
+    'NumpyArray',
+    'NumpyArraySize',
+    'NumpyBool',
+    'NumpyCountNonZero',
     'NumpyComplex',
     'NumpyComplex64',
     'NumpyComplex128',
+    'NumpyConjugate',
+    'NumpyEmpty',
+    'NumpyEmptyLike',
+    'NumpyFabs',
+    'NumpyFloat',
     'NumpyFloat32',
     'NumpyFloat64',
     'NumpyFull',
     'NumpyFullLike',
     'NumpyImag',
-    'NumpyBool',
+    'NumpyHypot',
     'NumpyInt',
     'NumpyInt8',
     'NumpyInt16',
@@ -84,11 +99,7 @@ __all__ = (
     'NumpyInt64',
     'NumpyLinspace',
     'NumpyMatmul',
-    'NumpyAmax',
-    'NumpyAmin',
-    'NumpyArange',
-    'NumpyArraySize',
-    'NumpyCountNonZero',
+    'NumpyNewArray',
     'NumpyMod',
     'NumpyNonZero',
     'NumpyNonZeroElement',
@@ -100,10 +111,11 @@ __all__ = (
     'NumpyRand',
     'NumpyRandint',
     'NumpyReal',
-    'Shape',
+    'NumpyTranspose',
     'NumpyWhere',
     'NumpyZeros',
     'NumpyZerosLike',
+    'Shape',
 )
 
 #=======================================================================================
@@ -188,12 +200,14 @@ class NumpyInt(PythonInt):
 class NumpyInt8(NumpyInt):
     """ Represents a call to numpy.int8() function.
     """
+    __slots__ = ()
     _precision = dtype_registry['int8'][1]
     name = 'int8'
 
 class NumpyInt16(NumpyInt):
     """ Represents a call to numpy.int16() function.
     """
+    __slots__ = ()
     _precision = dtype_registry['int16'][1]
     name = 'int16'
 
@@ -253,10 +267,12 @@ class NumpyImag(PythonImag):
     __slots__ = ('_precision','_rank','_shape','_order')
     name = 'imag'
[…]

[…]
