### 2023-01-12 pyccel/pyccel#1295 @boegel
Adding a `cat /proc/cpuinfo` and `cat /etc/os-release` + `uname -a` would already give a bunch of info, but still, I don't think GitHub Actions is suitable place to run benchmarks, it's no doubt subject to *a lot* of noise which could skew the results quite a bit.
