### 2023-01-12 pyccel/pyccel#1295 @EmilyBourne
I've found a machine to run on. I should be able to update the paper this evening
