+++ .github/workflows/Github_pytest.yml
@@ -89,20 +89,12 @@ jobs:
       - name: Parallel tests with pytest
         uses: ./.github/actions/pytest_parallel
 
-<<<<<<< HEAD
   Cuda:
-=======
-  CoverageChecker:
-
-    runs-on: ubuntu-latest
-    needs: Linux
->>>>>>> 26a14bfa9cae3e21b5175892b77f7bd1f7d334e2
 
     runs-on: ubuntu-20.04
     container: nvidia/cuda:11.7.1-devel-ubuntu20.04
     if: github.event.pull_request.base.ref != 'master'
     steps:
-<<<<<<< HEAD
       - uses: actions/checkout@v2
       - name: Prepare docker
         run: |
@@ -192,11 +184,6 @@ jobs:
       - uses: actions/checkout@v3
       - name: Set up Python 3.7
         uses: actions/setup-python@v4
-=======
-      - uses: actions/checkout@v3
-      - name: Set up Python 3.7
-        uses: actions/setup-python@v4
->>>>>>> 26a14bfa9cae3e21b5175892b77f7bd1f7d334e2
         with:
           python-version: 3.7
       - name: Install python dependencies
@@ -207,11 +194,7 @@ jobs:
       - name: Collect coverage information
         uses: actions/download-artifact@v3
         with:
-<<<<<<< HEAD
           name: coverage-artifact-xml
-=======
-          name: coverage-artifact
->>>>>>> 26a14bfa9cae3e21b5175892b77f7bd1f7d334e2
       - name: Collect diff information
         run: |
           BASE_BRANCH=$GITHUB_BASE_REF
