### 2023-01-12 f95382a
Update developer_docs/semantic_stage.md
Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>
