### 2023-01-12 1b1c9d7
Update developer_docs/semantic_stage.md
Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>
