### 2023-01-13 db8d76a
Revert "Put current context check in _assign_lhs_variable"
This reverts commit 4a8fab53276a878318e89d8ae32a5c22e114d8a4.
