+++ pyccel/parser/semantic.py
@@ -842,13 +842,17 @@ class SemanticParser(BasicParser):
                         severity = 'fatal')
 
             args, kwargs = split_positional_keyword_arguments(*args)
-
-            if func in (CupyArray, CudaArray):
+            if func in (CupyArray, CudaArray, NumpyArray):
                 if 'device' in self.scope.decorators or 'kernel' in self.scope.decorators:
                     current_context = 'device'
+                    if kwargs.get('memory_location', 'host') == 'host':
+                        errors.report("Host arrays cannot be allocated on the Device",
+                            symbol = expr,
+                            severity = 'fatal')
                 else:
                     current_context = 'host'
-                kwargs['current_context'] = current_context
+                if func in (CudaArray, CupyArray):
+                    kwargs['current_context'] = current_context
 
             for a in args:
                 if getattr(a,'dtype',None) == 'tuple':
@@ -1220,13 +1224,6 @@ class SemanticParser(BasicParser):
                     # Create Deallocate node
                     self._allocs[-1].append(lhs)
                 # ...
-
-                if lhs.is_ndarray:
-                    if 'device' in self.scope.decorators or 'kernel' in self.scope.decorators\
-                            and getattr(lhs, 'memory_location', 'host') == 'host':
-                        errors.report("Host arrays cannot be allocated on the Device",
-                            symbol = f"{lhs} = {rhs}",
-                            severity = 'fatal')
 
                 # We cannot allow the definition of a stack array in a loop
                 if lhs.is_stack_array and self.scope.is_loop:
