### 2023-01-13 865cc43
Also handle kwargs for built in functions
