### 2023-01-13 2c31a65
Don't handle lambdify in builtin_function
