### 2023-01-13 d086f55
Don't wrap symbolic objects
