+++ tests/symbolic/scripts/lambdas.py
@@ -12,3 +12,6 @@ m1 = lambdify(g1)
 print(f1)
 print(f2)
 print(g1)
+print(f1(2))
+print(g1(2))
+print(g1(2.0))
