### 2023-01-14 2896396
Remove old to_host / to_device test functions
