### 2023-01-14 3933de5
Add CupyArray/CudaArray to _infere_type
