### 2023-01-15 19c96db
Remove unused file
