+++ pyccel/ast/utilities.py
@@ -11,7 +11,6 @@ from itertools import chain
 from collections import namedtuple
 
 import pyccel.decorators as pyccel_decorators
-from pyccel.symbolic import lambdify
 from pyccel.errors.errors import Errors, PyccelError
 
 from .core          import (AsName, Import, FunctionDef, FunctionCall,
