### 2023-01-15 b9d6103
Remove unused decorator
