+++ pyccel/decorators.py
@@ -23,17 +23,6 @@ __all__ = (
     'types',
 )
 
-def lambdify(f):
-
-    args = f.__code__.co_varnames
-    from sympy import symbols
-    args = symbols(args)
-    expr = f(*args)
-    def wrapper(*vals):
-        return  expr.subs(zip(args,vals)).doit()
-
-    return wrapper
-
 def sympy(f):
     return f
 
