### 2023-01-16 #4991 @boegel
Update from my side: waiting for authors to come back with more reliable benchmark results - see https://github.com/pyccel/pyccel/issues/1295
