### 2023-01-17 #4991 @yguclu
@danielskatz We are running the benchmarks on two different supercomputing centers.

While the performance of Pyccel is in line with the test that we have run on GitHub Actions, the same is not true for Numba and Pythran (one is faster and the other much slower than before). Before presenting these results, we want to make sure that our configuration files are correct.
