### 2023-01-20 #4991 @yguclu
@danielskatz After further investigation we have realized that our tests are not completely adequate for a fair comparison with Numba and Pythran. Before publication I would like to address the following urgent points:

   1. Allow the user (or developer) to run any given test manually without having to copy lots of setup instructions from the script `run_benchmarks.py`;
   2. Make sure that every test returns a number or an array, which can be checked for correctness against the pure-Python version;
   3. Increase the run-time of the accelerated functions to at least a good fraction of a second (currently some of our tests terminate after less than 1 micro-second).

I kindly request a week of time for taking care of this.
