### 2023-01-20 #4991 @danielskatz
no worries @yguclu - please let us know when we should come back to this.  (and note that performance results are not strictly required for JOSS, but if they are there, they are part of the review)
