### 2023-01-31 #4991 @yguclu
The new benchmark suite is almost ready: see [this PR](https://github.com/pyccel/pyccel-benchmarks/pull/20).
