### 2023-01-31 #4991 @yguclu
@danielskatz @boegel We have just merged the aforementioned PR and we consider the benchmark suite as final. This is now our checklist for completing the article:

- [ ] Re-run the benchmarks (without the `-ffast-math` flag -- see [here](https://github.com/pyccel/pyccel-benchmarks/pull/17)) on a single dedicated node of our HPC facility
- [ ] Update the images in the manuscript
- [ ] Describe the architecture of the computing node in the manuscript (fixes [pyccel/pyccel/#1295](https://github.com/pyccel/pyccel/issues/1295))
- [ ] Update Pyccel version to 1.7.2 (and pin this version in the manuscript)
- [ ] Retrigger PDF build

We will check the items as we make progress. Please let me know if you see anything missing.
