### 2023-02-03 #4991 @yguclu
> @danielskatz @boegel We have just merged the aforementioned PR and we consider the benchmark suite as final. This is now our checklist for completing the article:
>
>     * [x]  Re-run the benchmarks (without the `-ffast-math` flag -- see [here](https://github.com/pyccel/pyccel-benchmarks/pull/17)) on a single dedicated node of our HPC facility
>
>     * [x]  Update the images in the manuscript
>
>     * [x]  Describe the architecture of the computing node in the manuscript (fixes [pyccel/pyccel/#1295](https://github.com/pyccel/pyccel/issues/1295))
>
>     * [x]  Update Pyccel version to 1.7.2 (and pin this version in the manuscript)
>
>     * [x]  Retrigger PDF build
>
>
> We will check the items as we make progress. Please let me know if you see anything missing.

Our checklist is complete. But the editorial header should now point to Pyccel version 1.7.2 instead of 1.7.0.
