### 2023-02-09 pyccel/pyccel#1295 @yguclu
@boegel Can we close this issue after PR [#20](https://github.com/pyccel/pyccel-benchmarks/pull/20)? As requested, the test results reported in the [pyccel/joss-paper](https://github.com/pyccel/pyccel/tree/joss-paper/joss-paper) branch do not use the `-ffast-math` flag.
