### 2023-03-09 #4991 @EmilyBourne
The tagged version is [v1.7.2](https://github.com/pyccel/pyccel/releases/tag/v1.7.2)
