### 2023-03-09 #4991 @yguclu
> The tagged version is [v1.7.2](https://github.com/pyccel/pyccel/releases/tag/v1.7.2)

Indeed. The latest Pyccel version on PyPI is now 1.7.3, but the benchmarks in the article were run with version 1.7.2. Hence we stick to 1.7.2.
