### 2023-03-09 #4991 @danielskatz
@yguclu and @EmilyBourne - I've suggested some minor changes in https://github.com/pyccel/pyccel/pull/1341 - please merge this, or let me know what you disagree with, then we can proceed to publication.
